XMonad.Actions.BluetileCommands

Description

This is a list of selected commands that can be made available using XMonad.Hooks.ServerMode to allow external programs to control the window manager. Bluetile (http://projects.haskell.org/bluetile/) uses this to enable its dock application to do things like changing workspaces and layouts.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

    import XMonad.Hooks.ServerMode
    import XMonad.Actions.BluetileCommands

Then edit your handleEventHook:

 main = xmonad defaultConfig { handleEventHook = serverModeEventHook' bluetileCommands }

See the documentation of XMonad.Hooks.ServerMode for details on how to actually invoke the commands from external programs.

bluetileCommands :: X [(String, X ())]