XMonad.Actions.CycleSelectedLayouts

Description

This module allows to cycle through the given subset of layouts.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad hiding ((|||))
 import XMonad.Layout.LayoutCombinators ((|||))
 import XMonad.Actions.CycleSelectedLayouts

   , ((modm,  xK_t ),   cycleThroughLayouts ["Tall", "Mirror Tall"])

Make sure you are using NewSelect from XMonad.Layout.LayoutCombinators, rather than the Select defined in xmonad core.

cycleThroughLayouts :: [String] -> X ()

If the current layout is in the list, cycle to the next layout. Otherwise, apply the first layout from list.