XMonad.Actions.DeManage

Description

This module provides a method to cease management of a window without unmapping it. This is especially useful for applications like kicker and gnome-panel. See also XMonad.Hooks.ManageDocks for more a more automated solution.

To make a panel display correctly with xmonad:

Determine the pixel size of the panel, add that value to defaultGaps

Launch the panel

Give the panel window focus, then press mod-d (or whatever key you have bound demanage to)

Convince the panel to move/resize to the correct location. Changing the panel's position setting several times seems to work.

Usage

To use demanage, add this import to your ~/.xmonad/xmonad.hs:

     import XMonad.Actions.DeManage

And add a keybinding, such as:

 , ((modm,               xK_d     ), withFocused demanage)

For detailed instructions on editing your key bindings, see XMonad.Doc.Extending.

demanage :: Window -> X ()

Stop managing the currently focused window.