XMonad.Actions.DwmPromote

Description

Dwm-like swap function for xmonad.

Swaps focused window with the master window. If focus is in the master, swap it with the next window in the stack. Focus stays in the master.

Usage

You can use this module with the following in your xmonad.hs:

   import XMonad.Actions.DwmPromote

then add a keybinding or substitute dwmpromote in place of promote:

  , ((modm,               xK_Return), dwmpromote)

For detailed instructions on editing your key bindings, see the tutorial.

dwmpromote :: X ()

Swap the focused window with the master window. If focus is in the master, swap it with the next window in the stack. Focus stays in the master.