XMonad.Actions.DynamicWorkspaceGroups

Description

Dynamically manage "workspace groups", sets of workspaces being used together for some common task or purpose, to allow switching between workspace groups in a single action. Note that this only makes sense for multi-head setups.

Usage

You can use this module by importing it into your ~/.xmonad/xmonad.hs file:

 import XMonad.Actions.DynamicWorkspaceGroups

Then add keybindings like the following (this example uses XMonad.Util.EZConfig-style keybindings, but this is not necessary):

    , ("M-y n", promptWSGroupAdd myXPConfig "Name this group: ")
    , ("M-y g", promptWSGroupView myXPConfig "Go to group: ")
    , ("M-y d", promptWSGroupForget myXPConfig "Forget group: ")

type WSGroupId = String

addWSGroup :: WSGroupId -> [WorkspaceId] -> X ()

Add a new workspace group with the given name.

addCurrentWSGroup :: WSGroupId -> X ()

Give a name to the current workspace group.

forgetWSGroup :: WSGroupId -> X ()

Delete the named workspace group from the list of workspace groups. Note that this has no effect on the workspaces involved; it simply forgets the given name.

viewWSGroup :: WSGroupId -> X ()

View the workspace group with the given name.

promptWSGroupView :: XPConfig -> String -> X ()

Prompt for a workspace group to view.

promptWSGroupAdd :: XPConfig -> String -> X ()

Prompt for a name for the current workspace group.

promptWSGroupForget :: XPConfig -> String -> X ()

Prompt for a workspace group to forget.

data WSGPrompt