XMonad.Actions.FlexibleResize

Description

Resize floating windows from any corner.

Usage

To use, first import this module into your ~/.xmonad/xmonad.hs file:

 import qualified XMonad.Actions.FlexibleResize as Flex

Then add an appropriate mouse binding:

     , ((modm, button3), (\w -> focus w >> Flex.mouseResizeWindow w))

For detailed instructions on editing your mouse bindings, see XMonad.Doc.Extending.

mouseResizeWindow

Arguments

:: Window
The window to resize.

-> X ()

Resize a floating window from whichever corner the mouse is closest to.

mouseResizeEdgeWindow

Arguments

:: Rational
The size of the area where only one edge is resized.

-> Window
The window to resize.

-> X ()

Resize a floating window from whichever corner or edge the mouse is closest to.