XMonad.Actions.PerLayoutKeys

Description

Define key-bindings on per-layout basis.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Actions.PerLayoutKeys

  ,((0, xK_F2), bindByLayout [("Tall", spawn "rxvt"), ("Mirror Tall", spawn "xeyes"), ("", spawn "xmessage hello")])

For detailed instructions on editing your key bindings, see XMonad.Doc.Extending.

chooseActionByLayout :: (String -> X ()) -> X ()

Uses supplied function to decide which action to run depending on current layout name.

bindByLayout :: [(String, X ())] -> X ()

If current layout is listed, run appropriate action (only the first match counts!) If it isn't listed, then run default action (marked with empty string, ""), or do nothing if default isn't supplied.