XMonad.Actions.ShowText

Description

ShowText displays text for sometime on the screen similar to XMonad.Util.Dzen which offers more features (currently)

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import XMonad.Actions.ShowText

Then add the event hook handler:

xmonad { handleEventHook = myHandleEventHooks <+> handleTimerEvent }

You can then use flashText in your keybindings:

((modMask, xK_Right), flashText def 1 "->" >> nextWS)

def :: Default a => a

defaultSTConfig :: ShowTextConfig

Deprecated: Use def (from Data.Default, and re-exported by XMonad.Actions.ShowText) instead.

handleTimerEvent :: Event -> X All

Handles timer events that notify when a window should be removed

flashText

Arguments

:: ShowTextConfig

-> Rational
number of seconds

-> String
text to display

-> X ()

Shows a window in the center of the screen with the given text

data ShowTextConfig

Constructors

STC

Fields

st_font :: String

Font name

st_bg :: String

Background color

st_fg :: String

Foreground color