XMonad.Actions.SimpleDate

Description

An example external contrib module for XMonad. Provides a simple binding to dzen2 to print the date as a popup menu.

Usage

To use, import this module into ~/.xmonad/xmonad.hs:

    import XMonad.Actions.SimpleDate

and add a keybinding, for example:

   , ((modm,               xK_d     ), date)

In this example, a popup date menu will now be bound to mod-d.

For detailed instructions on editing your key bindings, see XMonad.Doc.Extending.

date :: X ()