XMonad.Actions.WithAll

Description

Provides functions for performing a given action on all windows of the current workspace.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Actions.WithAll

then add a keybinding; for example:

, ((modm .|. shiftMask, xK_t), sinkAll)

For detailed instructions on editing your key bindings, see XMonad.Doc.Extending.

sinkAll :: X ()

Un-float all floating windows on the current workspace.

withAll :: (Window -> X ()) -> X ()

Execute an X action for each window on the current workspace.

withAll' :: (Window -> WindowSet -> WindowSet) -> X ()

Apply a function to all windows on the current workspace.

killAll :: X ()

Kill all the windows on the current workspace.