XMonad.Config.Azerty

Description

This module fixes some of the keybindings for the francophone among you who use an AZERTY keyboard layout. Config stolen from TeXitoi's config on the wiki.

Usage

To use this module, start with the following ~/.xmonad/xmonad.hs:

 import XMonad
 import XMonad.Config.Azerty

 main = xmonad azertyConfig

If you prefer, an azertyKeys function is provided which you can use as so:

 import qualified Data.Map as M
 main = xmonad someConfig { keys = \c -> azertyKeys c `M.union` keys someConfig c }

azertyConfig :: XConfig (Choose Tall (Choose (Mirror Tall) Full))

azertyKeys :: XConfig l -> Map (KeyMask, KeySym) (X ())