XMonad.Config.Bepo

Description

This module fixes some of the keybindings for the francophone among you who use a BEPO keyboard layout. Based on XMonad.Config.Azerty

Usage

To use this module, start with the following ~/.xmonad/xmonad.hs:

import XMonad
import XMonad.Config.Bepo

main = xmonad bepoConfig

If you prefer, an bepoKeys function is provided which you can use as so:

import qualified Data.Map as M
main = xmonad someConfig { keys = \c -> bepoKeys c `M.union` keys someConfig c }

bepoConfig :: XConfig (Choose Tall (Choose (Mirror Tall) Full))

bepoKeys :: XConfig l -> Map (KeyMask, KeySym) (X ())