XMonad.Config.Bluetile

Description

This is the default configuration of Bluetile (http://projects.haskell.org/bluetile/). If you are migrating from Bluetile to xmonad or want to create a similar setup, then this will give you pretty much the same thing, except for Bluetile's helper applications such as the dock.

Usage

To use this module, start with the following ~/.xmonad/xmonad.hs:

 import XMonad
 import XMonad.Config.Bluetile
 import XMonad.Util.Replace

 main = replace >> xmonad bluetileConfig

The invocation of replace will replace a currently running window manager. This is the default behaviour of Bluetile as well. See XMonad.Util.Replace for more information.

bluetileConfig :: XConfig (ModifiedLayout AvoidStruts (ModifiedLayout Minimize (ModifiedLayout BoringWindows (NewSelect (ModifiedLayout Rename (ModifiedLayout (Decoration ButtonDecoration DefaultShrinker) (ModifiedLayout Maximize (ModifiedLayout BorderResize PositionStoreFloat)))) (NewSelect (ModifiedLayout Rename (ModifiedLayout (Decoration WindowSwitcherDecoration DefaultShrinker) (ModifiedLayout DraggingVisualizer (ModifiedLayout Maximize MouseResizableTile)))) (NewSelect (ModifiedLayout Rename (ModifiedLayout (Decoration WindowSwitcherDecoration DefaultShrinker) (ModifiedLayout DraggingVisualizer (ModifiedLayout Maximize MouseResizableTile)))) (ModifiedLayout Rename (ModifiedLayout (Decoration WindowSwitcherDecoration DefaultShrinker) (ModifiedLayout DraggingVisualizer (ModifiedLayout Maximize (ModifiedLayout SmartBorder Full)))))))))))