XMonad.Config.Dmwit

Description

Deprecated: This module contains a personal configuration, to be removed from xmonad-contrib. If you use this module, please copy the relevant parts to your configuration or obtain a copy of it on https://xmonad.org/configurations.html and include it as a local module.

Documentation

outputOf :: String -> IO String

geomMean :: Floating a => [a] -> a

arithMean :: Floating a => [a] -> a

namedNumbers :: [Char] -> String -> [[Char]]

splitColon :: [Char] -> [[Char]]

parse :: (Floating a, Read a) => String -> a

modVolume :: String -> Integer -> IO Double

centerMouse :: X ()

statusBarMouse :: X ()

withScreen :: ScreenId -> (WorkspaceId -> WindowSet -> WindowSet) -> X ()

makeLauncher :: [Char] -> [Char] -> [Char] -> [Char] -> [Char]

launcher :: [Char]

termLauncher :: [Char]

viewShift :: (Ord a, Eq s, Eq i) => i -> StackSet i l a s sd -> StackSet i l a s sd

floatAll :: [String] -> Query (Endo WindowSet)

sinkFocus :: Ord a => StackSet i l a s sd -> StackSet i l a s sd

showMod :: String -> Integer -> X ()

volumeDzen :: String -> X ()

altMask :: KeyMask

bright :: String

dark :: String

fullscreen43on169 :: RationalRect

fullscreenMPlayer :: Query (Endo (StackSet PhysicalWorkspace (Layout Window) Window ScreenId ScreenDetail))

operationOn :: (Ord t, Eq s) => (t -> StackSet PhysicalWorkspace l t s sd -> StackSet PhysicalWorkspace l t s sd) -> ScreenId -> VirtualWorkspace -> t -> Query (Endo (StackSet PhysicalWorkspace l t s sd))

viewFullOn :: (Ord t, Eq s) => ScreenId -> VirtualWorkspace -> t -> Query (Endo (StackSet PhysicalWorkspace l t s sd))

centerWineOn :: (Ord t, Eq s) => ScreenId -> VirtualWorkspace -> t -> Query (Endo (StackSet PhysicalWorkspace l t s sd))

class Show a => PPrint a where

Minimal complete definition

Nothing

Methods

pprint :: Int -> a -> String

data PPrintable

Constructors

forall a.PPrint a => P a

record :: String -> Int -> [(String, PPrintable)] -> String

dmwitConfig :: ScreenId -> XConfig (ModifiedLayout Magnifier (Choose (ModifiedLayout AvoidStruts Grid) (ModifiedLayout WithBorder Full)))

main :: IO ()

keyBindings :: XConfig Layout -> Map (KeyMask, KeySym) (X ())

atSchool :: MonadIO m => b -> b -> m b

anyMask :: [((KeyMask, b), b)] -> [((KeyMask, b), b)]

pipeName :: Show a => [Char] -> a -> [Char]

xmobarCommand :: ScreenId -> String

allPPs :: ScreenId -> X ()

color :: String -> String -> String

ppFocus :: ScreenId -> PP

ppWorkspaces :: ScreenId -> PP