XMonad.Config.Kde

Description

This module provides a config suitable for use with the KDE desktop environment.

Usage

To use this module, start with the following ~/.xmonad/xmonad.hs:

 import XMonad
 import XMonad.Config.Kde

 main = xmonad kdeConfig

For KDE 4, replace kdeConfig with kde4Config

For examples of how to further customize kdeConfig see XMonad.Config.Desktop.

kdeConfig :: XConfig (ModifiedLayout AvoidStruts (Choose Tall (Choose (Mirror Tall) Full)))

kde4Config :: XConfig (ModifiedLayout AvoidStruts (Choose Tall (Choose (Mirror Tall) Full)))