XMonad.Config.Xfce

Description

This module provides a config suitable for use with the Xfce desktop environment.

Usage

To use this module, start with the following ~/.xmonad/xmonad.hs:

 import XMonad
 import XMonad.Config.Xfce

 main = xmonad xfceConfig

For examples of how to further customize xfceConfig see XMonad.Config.Desktop.

xfceConfig :: XConfig (ModifiedLayout AvoidStruts (Choose Tall (Choose (Mirror Tall) Full)))