XMonad.Hooks.CurrentWorkspaceOnTop

Description

Ensures that the windows of the current workspace are always in front of windows that are located on other visible screens. This becomes important if you use decoration and drag windows from one screen to another. Using this module, the dragged window will always be in front of other windows.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Hooks.CurrentWorkspaceOnTop

 main = xmonad $ defaultConfig {
    ...
    logHook = currentWorkspaceOnTop
    ...
  }

currentWorkspaceOnTop :: X ()