XMonad.Hooks.DebugStack

Description

Dump the state of the StackSet. A logHook and handleEventHook are also provided.

Documentation

debugStack :: X ()

Print the state of the current window stack for the current workspace to stderr, which for most installations goes to ~/.xsession-errors. XMonad.Util.DebugWindow is used to display the individual windows.

debugStackFull :: X ()

Print the state of the current window stack for all workspaces to stderr, which for most installations goes to ~/.xsession-errors. XMonad.Util.DebugWindow is used to display the individual windows.

debugStackString :: X String

Dump the state of the current workspace in the StackSet as a multiline String.

debugStackFullString :: X String

Dump the state of all workspaces in the StackSet as a multiline String. @@@ this is in stackset order, which is roughly lru-ish

debugStackLogHook :: X ()

debugStack packaged as a logHook. (Currently this is identical.)

debugStackFullLogHook :: X ()

'debugStackFull packaged as a logHook. (Currently this is identical.)

debugStackEventHook :: Event -> X All

debugStack packaged as a handleEventHook. You almost certainly do not want to use this unconditionally, as it will cause massive amounts of output and possibly slow xmonad down severely.

debugStackFullEventHook :: Event -> X All

debugStackFull packaged as a handleEventHook. You almost certainly do not want to use this unconditionally, as it will cause massive amounts of output and possibly slow xmonad down severely.