XMonad.Hooks.ICCCMFocus

Description

Deprecated: XMonad.Hooks.ICCCMFocus: xmonad>0.10 core merged issue 177

Implemented in your logHook, Java swing applications will not misbehave when it comes to taking and losing focus.

This has been done by taking the patch in http://code.google.com/p/xmonad/issues/detail?id=177 and refactoring it so that it can be included in ~/.xmonad/xmonad.hs.

   conf' =
     conf {
       logHook = takeTopFocus
     }

Documentation

atom_WM_TAKE_FOCUS :: X Atom

takeFocusX :: Window -> X ()

takeTopFocus :: X ()

The value to add to your log hook configuration.