XMonad.Hooks.InsertPosition

Description

Configure where new windows should be added and which window should be focused.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Hooks.InsertPosition
 xmonad defaultConfig { manageHook = insertPosition Master Newer <+> myManageHook }

You should you put the manageHooks that use doShift to take effect before insertPosition, so that the window order will be consistent. Because ManageHooks compose from right to left (like function composition .), this means that insertPosition should be the leftmost ManageHook.

insertPosition :: Position -> Focus -> ManageHook

insertPosition. A manage hook for placing new windows. XMonad's default is the same as using: insertPosition Above Newer.

data Focus

Constructors

Newer

Older

data Position

Constructors

Master

End

Above

Below