XMonad.Hooks.Minimize

Description

Handles window manager hints to minimize and restore windows. Use this with XMonad.Layout.Minimize.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Hooks.Minimize
 import XMonad.Layout.Minimize

 myHandleEventHook = minimizeEventHook
 myLayout = minimize (Tall 1 (3/100) (1/2)) ||| Full ||| etc..
 main = xmonad defaultConfig { layoutHook = myLayout
                             , handleEventHook = myHandleEventHook }

minimizeEventHook :: Event -> X All