XMonad.Hooks.Rescreen

Description

Custom hooks for screen (xrandr) configuration changes.

Usage

This module provides a replacement for the screen configuration change handling in core that enables attaching custom hooks to screen (xrandr) configuration change events. These can be used to restart/reposition status bars or systrays automatically after xrandr, as well as to actually invoke xrandr or autorandr when an output is (dis)connected.

You can use it by including the following in your ~/.xmonad/xmonad.hs:

import XMonad.Hooks.RescreenHook

defining your custom hooks:

myAfterRescreenHook :: X ()
myAfterRescreenHook = …

myRandrChangeHook :: X ()
myRandrChangeHook = …

rescreenCfg = def{
    afterRescreenHook = myAfterRescreenHook,
    randrChangeHook = myRandrChangeHook
}

and adding rescreenHook to your xmonad config:

main = xmonad $ … . rescreenHook rescreenCfg . … $ def{…}

data RescreenConfig

Hook configuration for rescreenHook.

Constructors

RescreenConfig

Fields

afterRescreenHook :: X ()

hook to invoke after rescreen

randrChangeHook :: X ()

hook for other randr changes, e.g. (dis)connects

addAfterRescreenHook :: X () -> XConfig l -> XConfig l

Shortcut for rescreenHook.

addRandrChangeHook :: X () -> XConfig l -> XConfig l

Shortcut for rescreenHook.

rescreenHook :: RescreenConfig -> XConfig l -> XConfig l

Attach custom hooks to screen (xrandr) configuration change events. Replaces the built-in rescreen handling of xmonad core with:

listen to RRScreenChangeNotifyEvent in addition to ConfigureEvent on the root window

whenever such event is received:

clear any other similar events (Xorg server emits them in bunches)

if any event was ConfigureEvent, rescreen and invoke afterRescreenHook

if there was no ConfigureEvent, invoke randrChangeHook only

afterRescreenHook is useful for restarting/repositioning status bars and systray.

randrChangeHook may be used to automatically trigger xrandr (or perhaps autorandr) when outputs are (dis)connected.

Note that rescreenHook is safe to use several times, rescreen is still done just once and hooks are invoked in sequence, also just once.