XMonad.Hooks.ScreenCorners

Description

Run X () actions by touching the edge of your screen with your mouse.

Usage

This extension adds KDE-like screen corners to XMonad. By moving your cursor into one of your screen corners you can trigger an X () action, for example XMonad.Actions.GridSelect.goToSelected or XMonad.Actions.CycleWS.nextWS etc.

To use it, import it on top of your xmonad.hs:

 import XMonad.Hooks.ScreenCorners

Then add your screen corners in our startup hook:

 myStartupHook = do
     ...
     addScreenCorner SCUpperRight (goToSelected defaultGSConfig { gs_cellwidth = 200})
     addScreenCorners [ (SCLowerRight, nextWS)
                      , (SCLowerLeft,  prevWS)
                      ]

And finally wait for screen corner events in your event hook:

 myEventHook e = do
     ...
     screenCornerEventHook e

Adding screen corners

data ScreenCorner

Constructors

SCUpperLeft

SCUpperRight

SCLowerLeft

SCLowerRight

addScreenCorner :: ScreenCorner -> X () -> X ()

Add one single X () action to a screen corner

addScreenCorners :: [(ScreenCorner, X ())] -> X ()

Add a list of (ScreenCorner, X ()) tuples

Event hook

screenCornerEventHook :: Event -> X All

Handle screen corner events