XMonad.Hooks.ShowWName

Description

Flash the names of workspaces name when switching to them. This is a reimplementation of ShowWName as a logHook.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import XMonad.Hooks.ShowWName

main :: IO ()
main = xmonad $ def
  { logHook = showWNameLogHook def
  }

Whenever a workspace gains focus, the above logHook will flash its name. You can customise the duration of the flash, as well as colours by customising the SWNConfig argument that showWNameLogHook takes.

Alternatively, you can also bind flashName to a key and manually invoke it when you want to know which workspace you are on.

showWNameLogHook :: SWNConfig -> X ()

LogHook for flashing the name of a workspace upon entering it.

data SWNConfig

Constructors

SWNC

Fields

swn_font :: String

Font name

swn_bgcolor :: String

Background color

swn_color :: String

String color

swn_fade :: Rational

Time in seconds of the name visibility

flashName :: SWNConfig -> X ()

Flash the name of the currently focused workspace.