XMonad.Hooks.TaffybarPagerHints

Description

This module exports additional X properties that allow taffybar to understand the state of XMonad.

Documentation

You can use this module with the following in your xmonad.hs file:

import XMonad.Hooks.TaffybarPagerHints (pagerHints)

main = xmonad $ ewmh $ pagerHints $ def
...

pagerHints :: XConfig a -> XConfig a

Add support for the "Current Layout" and "Visible Workspaces" custom hints to the given config.

pagerHintsLogHook :: X ()

Update the current values of both custom hints.

pagerHintsEventHook :: Event -> X All

Handle all "Current Layout" events received from pager widgets, and set the current layout accordingly.

setCurrentLayoutProp :: String -> X ()

Set the value of the "Current Layout" custom hint to the one given.

setVisibleWorkspacesProp :: [String] -> X ()

Set the value of the "Visible Workspaces" hint to the one given.