XMonad.Hooks.WorkspaceByPos

Description

Useful in a dual-head setup: Looks at the requested geometry of new windows and moves them to the workspace of the non-focused screen if necessary.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import XMonad.Hooks.WorkspaceByPos

myManageHook = workspaceByPos <+> manageHook def

main = xmonad def { manageHook = myManageHook }

workspaceByPos :: ManageHook