XMonad.Layout.Accordion

Description

LayoutClass that puts non-focused windows in ribbons at the top and bottom of the screen.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import XMonad.Layout.Accordion

Then edit your layoutHook by adding the Accordion layout:

myLayout = Accordion ||| Full ||| etc..
main = xmonad def { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

data Accordion a

Constructors

Accordion