XMonad.Layout.BorderResize

Description

This layout modifier will allow to resize windows by dragging their borders with the mouse. However, it only works in layouts or modified layouts that react to the SetGeometry message. XMonad.Layout.WindowArranger can be used to create such a setup, but it is probably must useful in a floating layout such as XMonad.Layout.PositionStoreFloat with which it has been mainly tested. See the documentation of PositionStoreFloat for a typical usage example.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import XMonad.Layout.BorderResize
myLayout = borderResize (... layout setup that reacts to SetGeometry ...)
main = xmonad def { layoutHook = myLayout }

borderResize :: l a -> ModifiedLayout BorderResize l a

data BorderResize a

Constructors

BR (Map Window RectWithBorders)

type RectWithBorders = (Rectangle, [BorderInfo])

data BorderInfo