XMonad.Layout.ButtonDecoration

Description

A decoration that includes small buttons on both ends which invoke various actions when clicked on: Show a window menu (see XMonad.Actions.WindowMenu), minimize, maximize or close the window.

Note: For maximizing and minimizing to actually work, you will need to integrate XMonad.Layout.Maximize and XMonad.Layout.Minimize into your setup. See the documentation of those modules for more information.

Usage:

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import XMonad.Layout.DecorationAddons
import XMonad.Layout.ButtonDecoration

Then edit your layoutHook by adding the ButtonDecoration to your layout:

myL = buttonDeco shrinkText defaultThemeWithButtons (layoutHook def)
main = xmonad def { layoutHook = myL }

buttonDeco :: (Eq a, Shrinker s) => s -> Theme -> l a -> ModifiedLayout (Decoration ButtonDecoration s) l a

data ButtonDecoration a