XMonad.Layout.CenteredIfSingle

Description

A layout modifier that, if there is only a single window on screen, places that window in the center of the screen. This is especially useful on wide screen setups, where the window would otherwise be unnecessarily far away from the center of your field of vision.

Usage

You can use this module by including the following in your xmonad.hs:

import XMonad.Layout.CenteredIfSingle

and adding the centeredIfSingle layoutmodifier to your layouts.

myLayoutHook = centeredIfSingle 0.7 0.8 Grid ||| ...

For more information on configuring your layouts see the tutorial and XMonad.Doc.Extending.

centeredIfSingle

Arguments

:: Double
Horizontal ratio of the screen the centered window should take up; should be a value between 0.0 and 1.0

-> Double
Vertical ratio; should also be a value between 0.0 and 1.0

-> l a
The layout that will be used if more than one window is open

-> ModifiedLayout CenteredIfSingle l a

Layout Modifier that places a window in the center of the screen, leaving room on all sides if there is only a single window

data CenteredIfSingle a

Layout Modifier that places a window in the center of the screen, leaving room on the left and right if there is only a single window. The first argument is the horizontal and the second one the vertical ratio of the screen the centered window should take up. Both numbers should be between 0.0 and 1.0.