XMonad.Layout.Circle

Description

Circle is an elliptical, overlapping layout, by Peter De Wachter

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Layout.Circle

Then edit your layoutHook by adding the Circle layout:

 myLayout = Circle ||| Full ||| etc..
 main = xmonad defaultConfig { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

data Circle a

Constructors

Circle