XMonad.Layout.DecorationEx.DwmGeometry

Description

This defines window decorations which are shown as a bar of fixed width on top of window.

Usage:

You can use this module with the following in your xmonad.hs:

import XMonad.Layout.DecorationEx.DwmStyle

Then edit your layoutHook by adding the DwmStyle decoration to your layout:

myL = dwmStyleDeco shrinkText (layoutHook def)
main = xmonad def { layoutHook = myL }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

data DwmGeometry a

Decoration geometry data type

Constructors

DwmGeometry

Fields

dwmShowForFocused :: !Bool

Whether to show decorations on focused windows

dwmHorizontalPosition :: !Rational

Horizontal position of decoration rectangle. 0 means place it at left corner, 1 - place it at right corner, 1%2 - place it at center.

dwmDecoHeight :: !Dimension

Height of decoration rectangle

dwmDecoWidth :: !Dimension

Width of decoration rectangle

dwmStyleDeco

Arguments

:: Shrinker shrinker

=> shrinker
Strings shrinker, for example shrinkText

-> ThemeEx StandardWidget
Decoration theme (font, colors, widgets, etc)

-> l Window
Layout to be decorated

-> ModifiedLayout (DecorationEx TextDecoration StandardWidget DwmGeometry shrinker) l Window

Add a decoration to window layout. Widgets are indicated with text fragments using TextDecoration; decoration placement is similar to DWM.

dwmStyleDecoEx

Arguments

:: Shrinker shrinker

=> shrinker
Strings shrinker, for example shrinkText

-> DwmGeometry Window

-> ThemeEx StandardWidget
Decoration theme (font, colors, widgets, etc)

-> l Window
Layout to be decorated

-> ModifiedLayout (DecorationEx TextDecoration StandardWidget DwmGeometry shrinker) l Window

Add a decoration to window layout. Widgets are indicated with text fragments using TextDecoration; decoration placement can be adjusted.