XMonad.Layout.DecorationEx.Geometry

Description

This module defines DecorationGeometry type class, and default implementation for it.

Documentation

class (Read (geom a), Show (geom a), Eq a) => DecorationGeometry geom a where

Decoration geometry class. Decoration geometry is responsible for placement of window decorations: whether they should be on the top of the window or on the bottom, should they go for full window width or only be of certain width, etc. This does not know what will be drawn inside decorations.

Minimal complete definition

describeGeometry, pureDecoration

Methods

describeGeometry :: geom a -> String

Give a name to decoration geometry implementation.

shrinkWindow :: geom a -> Rectangle -> Rectangle -> Rectangle

Reduce original window size to make space for decoration, if necessary.

pureDecoration

Arguments

:: geom a
Decoration geometry instance

-> Rectangle
Screen rectangle

-> Stack a
Current stack of windows being displayed

-> [(a, Rectangle)]
Set of all windows with their corresponding rectangle

-> (a, Rectangle)
Window being decorated and its rectangle

-> Maybe Rectangle

The pure version of the main method, decorate. The method should return a rectangle where to place window decoration, or Nothing if this window is not to be decorated.

decorateWindow

Arguments

:: geom a
Decoration geometry instance

-> Rectangle
Screen rectangle

-> Stack a
Current stack of windows being displayed

-> [(a, Rectangle)]
Set of all windows with their corresponding rectangle

-> (a, Rectangle)
Window being decorated and its rectangle

-> X (Maybe Rectangle)

The method should return a rectangle where to place window decoration, or Nothing if this window is not to be decorated.

newtype DefaultGeometry a

Data type for default implementation of DecorationGeometry. This defines simple decorations: a horizontal bar at the top of each window, running for full width of the window.

Constructors

DefaultGeometry

Fields

gDecorationHeight :: Dimension