XMonad.Layout.DecorationEx.LayoutModifier

Description

Layout modifier, which is responsible for creation of decoration rectangles (windows), updating and removing them when needed. It is parameterized by DecorationGeometry, which says where decorations should be placed, and by DecorationEngine, which says how decorations should look.

Usage

This module exports decorationEx function, which is a generic function for adding decorations to your layouts. It can be used to use different decoration geometries and engines in any combination. For most used combinations, there are convenience functions in XMonad.Layout.DecorationEx.TextEngine, XMonad.Layout.DecorationEx.TabbedGeometry, and XMonad.Layout.DecorationEx.DwmGeometry.

You can use this module with the following in your xmonad.hs:

import XMonad.Layout.DecorationEx.LayoutModifier

Then edit your layoutHook by adding the DwmStyle decoration to your layout:

myL = decorationEx shrinkText myTheme myEngine myGeometry (layoutHook def)
        where
          myGeometry = DefaultGeometry -- or another geometry type
          myEngine = TextDecoration    -- or another decoration engine
          myTheme = GenericTheme {...} -- theme type should correspond to selected engine type

main = xmonad def { layoutHook = myL }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

decorationEx

Arguments

:: (DecorationEngine engine widget a, DecorationGeometry geom a, Shrinker shrinker)

=> shrinker
Strings shrinker, for example shrinkText

-> Theme engine widget
Decoration theme

-> engine widget a
Decoration engine instance

-> geom a
Decoration geometry instance

-> l a
Underlying layout to be decorated

-> ModifiedLayout (DecorationEx engine widget geom shrinker) l a

Apply a DecorationEx modifier to an underlying layout

data DecorationEx engine widget geom shrinker a

The DecorationEx LayoutModifier. This data type is an instance of the LayoutModifier class. This data type will be passed, together with a layout, to the ModifiedLayout type constructor to modify the layout by adding decorations according to a DecorationEngine.