XMonad.Layout.DecorationEx.TabbedGeometry

Description

This module defines window decoration geometry based on tabs. The tabs can follow horizontally and be placed above or below windows; in such case, tabs can occupy full width of the window or be aligned to left or right. Or tabs can go vertically near left or right side of the window.

Documentation

textTabbed

Arguments

:: Shrinker shrinker

=> shrinker
Strings shrinker, e.g. shrinkText

-> ThemeEx StandardWidget
Decoration theme

-> l Window
Layout to be decorated

-> ModifiedLayout (DecorationEx TextDecoration StandardWidget TabbedGeometry shrinker) l Window

Add tabbed decorations (with default settings) with text-based widgets to a layout.

data TabbedGeometry a

Constructors

HorizontalTabs

Fields

showIfSingleWindow :: !SingleTabMode

What to do if there is only one tab

hTabPlacement :: !HorizontalTabPlacement

Where to place horizontal tabs

hTabAlignment :: !HorizontalTabsAlignment

How to align horizontal tabs (makes sense with fixed width of tabs).

hTabWidth :: !HorizontalTabWidth

Width of horizontal tabs

hTabHeight :: !Dimension

Height of horizontal tabs

VerticalTabs

Fields

showIfSingleWindow :: !SingleTabMode

What to do if there is only one tab

vTabPlacement :: !VerticalTabPlacement

Where to place vertical tabs

vTabWidth :: !Dimension

Width of vertical tabs

vTabHeight :: !Dimension

Height of vertical tabs

data HorizontalTabPlacement

Placement of tabs when they go horizontally: should they be placed above or below the window.

Constructors

Top

Bottom

data VerticalTabPlacement

Placement of tabs when they go vertically: should they appear at left or at right side of the window.

Constructors

TabsAtLeft

TabsAtRight

data HorizontalTabWidth

Width of tabs when they go horizontally.

Constructors

AutoWidth
Define the width automatically by evenly dividing windows' width

FixedWidth !Dimension
Use fixed width of the tab

data HorizontalTabsAlignment

Alignment of tabs when they go horizontally.

Constructors

AlignTabsLeft

AlignTabsCenter

AlignTabsRight

data SingleTabMode

What to do if there is only one tab.

Constructors

ShowTab

HideTab