XMonad.Layout.DecorationEx.TextEngine

Description

Window decoration engine, that uses text fragments (like "[X]") to indicate widgets (window buttons).

Documentation

textDecoration

Arguments

:: Shrinker shrinker

=> shrinker
String shrinker, for example shrinkText

-> Theme TextDecoration StandardWidget
Decoration theme (font, colors, widgets, etc)

-> l Window
Layout to be decorated

-> ModifiedLayout (DecorationEx TextDecoration StandardWidget DefaultGeometry shrinker) l Window

Add decoration to existing layout. Widgets are indicated by text fragments, like "[+]". Geometry is simple: a horizontal panel at the top of each window, going for the full width of the window.

data TextDecoration widget a

Decoration engine data type

Constructors

TextDecoration