XMonad.Layout.DecorationEx.Widgets

Description

This module contains data types and utilities to deal with decoration widgets. A widget is anything that is displayed on window decoration, and, optionally, can react on clicks. Examples of widgets are usual window buttons (minimize, maximize, close), window icon and window title.

Data types

data StandardCommand

Standard window commands.

One can extend this list by simply doing

data MyWindowCommand =
    Std StandardCommand
  | SomeFancyCommand

instance WindowCommand MyWindowCommand where ...

type MyWidget = GenericWidget MyWindowCommand

Constructors

FocusWindow
Focus the window

FocusUp
Move focus to previous window

FocusDown
Move focus to following window

MoveToNextGroup
Move the window to the next group (see XMonad.Layout.Groups)

MoveToPrevGroup
Move the window to the previous group

DwmPromote
Execute dwmpromote (see XMonad.Actions.DwmPromote)

ToggleSticky
Make window sticky or unstick it (see XMonad.Actions.CopyWindow)

ToggleMaximize
Maximize or restore window (see XMonad.Layout.Maximize)

Minimize
Minimize window (see XMonad.Actions.Minimize)

CloseWindow
Close the window

GridWindowMenu
Show window menu via XMonad.Actions.GridSelect (see XMonad.Actions.WindowMenu)

class DecorationWidget widget => TextWidget widget where

Type class for widgets that can be displayed as text fragments by TextDecoration engine.

Methods

widgetString :: DrawData engine widget -> widget -> X String

data GenericWidget cmd

Generic data type for decoration widgets.

Constructors

TitleWidget
Window title (just text label)

WindowIcon
Window icon with some associated command | Other widgets

Fields

swCommand :: !cmd

Window command

GenericWidget

Fields

swCheckedText :: !String

Text for checked widget state

swUncheckedText :: !String

Text for unchecked widget state

swCommand :: !cmd

Window command

type StandardWidget = GenericWidget StandardCommand

Generic widget type specialized for StandardCommand

Utility functions

isWidgetChecked :: DecorationWidget widget => widget -> Window -> X Bool

Check if the widget should be displayed in checked state.

Presets for standard widgets

titleW :: StandardWidget

Widget for window title

toggleStickyW :: StandardWidget

Widget for ToggleSticky command.

minimizeW :: StandardWidget

Widget for Minimize command

maximizeW :: StandardWidget

Widget for ToggleMaximize command

closeW :: StandardWidget

Widget for CloseWindow command

dwmpromoteW :: StandardWidget

moveToNextGroupW :: StandardWidget

moveToPrevGroupW :: StandardWidget