XMonad.Layout.DecorationMadness

Description

A collection of decorated layouts: some of them may be nice, some usable, others just funny.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Layout.DecorationMadness

Then edit your layoutHook by adding the layout you want:

 main = xmonad defaultConfig { layoutHook = someMadLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

You can also edit the default theme:

 myTheme = defaultTheme { inactiveBorderColor = "#FF0000"
                        , activeTextColor     = "#00FF00" }

and

 mylayout = tabbed shrinkText myTheme ||| Full ||| etc..

When a layout is resizable, this means two different things: you can grab a window's decoration with the pointer and move it around, and you can move and resize windows with the keyboard. For setting up the key bindings, please read the documentation of XMonad.Layout.WindowArranger

The default theme can be dynamically change with the xmonad theme selector. See XMonad.Prompt.Theme. For more themes, look at XMonad.Util.Themes

Decorated layouts based on Circle

Here you will find Circle based decorated layouts.

circleSimpleDefault :: ModifiedLayout (Decoration DefaultDecoration DefaultShrinker) Circle Window

A Circle layout with the xmonad default decoration, default theme and default shrinker.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/circleSimpleDefault.png

circleDefault :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration DefaultDecoration s) Circle Window

Similar to circleSimpleDefault but with the possibility of setting a custom shrinker and a custom theme.

circleSimpleDefaultResizable :: ModifiedLayout (Decoration DefaultDecoration DefaultShrinker) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger Circle)) Window

A Circle layout with the xmonad default decoration, default theme and default shrinker, but with the possibility of moving windows with the mouse, and resize/move them with the keyboard.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/circleSimpleDefaultResizable.png

circleDefaultResizable :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration DefaultDecoration s) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger Circle)) Window

Similar to circleSimpleDefaultResizable but with the possibility of setting a custom shrinker and a custom theme.

circleSimpleDeco :: ModifiedLayout (Decoration SimpleDecoration DefaultShrinker) Circle Window

A Circle layout with the xmonad simple decoration, default theme and default shrinker.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/circleSimpleDeco.png

circleSimpleDecoResizable :: ModifiedLayout (Decoration SimpleDecoration DefaultShrinker) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger Circle)) Window

A Circle layout with the xmonad simple decoration, default theme and default shrinker, but with the possibility of moving windows with the mouse, and resize/move them with the keyboard.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/circleSimpleDecoResizable.png

circleDeco :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration SimpleDecoration s) Circle Window

Similar to circleSimpleDece but with the possibility of setting a custom shrinker and a custom theme.

circleDecoResizable :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration SimpleDecoration s) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger Circle)) Window

Similar to circleSimpleDecoResizable but with the possibility of setting a custom shrinker and a custom theme.

circleSimpleDwmStyle :: ModifiedLayout (Decoration DwmStyle DefaultShrinker) Circle Window

A Circle layout with the xmonad DwmStyle decoration, default theme and default shrinker.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/circleSimpleDwmStyle.png

circleDwmStyle :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration DwmStyle s) Circle Window

Similar to circleSimpleDwmStyle but with the possibility of setting a custom shrinker and a custom theme.

circleSimpleTabbed :: ModifiedLayout (Decoration TabBarDecoration DefaultShrinker) (ModifiedLayout ResizeScreen Circle) Window

A Circle layout with the xmonad tabbed decoration, default theme and default shrinker.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/circleSimpleTabbed.png

circleTabbed :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration TabBarDecoration s) (ModifiedLayout ResizeScreen Circle) Window

Similar to circleSimpleTabbed but with the possibility of setting a custom shrinker and a custom theme.

Decorated layouts based on Accordion

Here you will find decorated layouts based on the Accordion layout.

accordionSimpleDefault :: ModifiedLayout (Decoration DefaultDecoration DefaultShrinker) Accordion Window

An Accordion layout with the xmonad default decoration, default theme and default shrinker.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/accordionSimpleDefault.png

accordionDefault :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration DefaultDecoration s) Accordion Window

Similar to accordionSimpleDefault but with the possibility of setting a custom shrinker and a custom theme.

accordionSimpleDefaultResizable :: ModifiedLayout (Decoration DefaultDecoration DefaultShrinker) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger Accordion)) Window

An Accordion layout with the xmonad default decoration, default theme and default shrinker, but with the possibility of moving windows with the mouse, and resize/move them with the keyboard.

accordionDefaultResizable :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration DefaultDecoration s) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger Accordion)) Window

Similar to accordionSimpleDefaultResizable but with the possibility of setting a custom shrinker and a custom theme.

accordionSimpleDeco :: ModifiedLayout (Decoration SimpleDecoration DefaultShrinker) Accordion Window

An Accordion layout with the xmonad simple decoration, default theme and default shrinker.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/accordionSimpleDeco.png

accordionSimpleDecoResizable :: ModifiedLayout (Decoration SimpleDecoration DefaultShrinker) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger Accordion)) Window

An Accordion layout with the xmonad simple decoration, default theme and default shrinker, but with the possibility of moving windows with the mouse, and resize/move them with the keyboard.

accordionDeco :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration SimpleDecoration s) Accordion Window

Similar to accordionSimpleDece but with the possibility of setting a custom shrinker and a custom theme.

accordionDecoResizable :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration SimpleDecoration s) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger Accordion)) Window

Similar to accordionSimpleDecoResizable but with the possibility of setting a custom shrinker and a custom theme.

accordionSimpleDwmStyle :: ModifiedLayout (Decoration DwmStyle DefaultShrinker) Accordion Window

An Accordion layout with the xmonad DwmStyle decoration, default theme and default shrinker.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/accordionSimpleDwmStyle.png

accordionDwmStyle :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration DwmStyle s) Accordion Window

Similar to accordionSimpleDwmStyle but with the possibility of setting a custom shrinker and a custom theme.

accordionSimpleTabbed :: ModifiedLayout (Decoration TabBarDecoration DefaultShrinker) (ModifiedLayout ResizeScreen Accordion) Window

An Accordion layout with the xmonad tabbed decoration, default theme and default shrinker.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/accordionSimpleTabbed.png

accordionTabbed :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration TabBarDecoration s) (ModifiedLayout ResizeScreen Accordion) Window

Similar to accordionSimpleTabbed but with the possibility of setting a custom shrinker and a custom theme.

Tall decorated layouts

In this section you will find decorated layouts based on the Tall layout.

tallSimpleDefault :: ModifiedLayout (Decoration DefaultDecoration DefaultShrinker) Tall Window

A Tall layout with the xmonad default decoration, default theme and default shrinker.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/tallSimpleDefault.png

tallDefault :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration DefaultDecoration s) Tall Window

Similar to tallSimpleDefault but with the possibility of setting a custom shrinker and a custom theme.

tallSimpleDefaultResizable :: ModifiedLayout (Decoration DefaultDecoration DefaultShrinker) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger Tall)) Window

A Tall layout with the xmonad default decoration, default theme and default shrinker, but with the possibility of moving windows with the mouse, and resize/move them with the keyboard.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/tallSimpleDefaultResizable.png

tallDefaultResizable :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration DefaultDecoration s) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger Tall)) Window

Similar to tallSimpleDefaultResizable but with the possibility of setting a custom shrinker and a custom theme.

tallSimpleDeco :: ModifiedLayout (Decoration SimpleDecoration DefaultShrinker) Tall Window

A Tall layout with the xmonad simple decoration, default theme and default shrinker.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/tallSimpleDeco.png

tallDeco :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration SimpleDecoration s) Tall Window

Similar to tallSimpleDece but with the possibility of setting a custom shrinker and a custom theme.

tallSimpleDecoResizable :: ModifiedLayout (Decoration SimpleDecoration DefaultShrinker) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger Tall)) Window

A Tall layout with the xmonad simple decoration, default theme and default shrinker, but with the possibility of moving windows with the mouse, and resize/move them with the keyboard.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/tallSimpleDecoResizable.png

tallDecoResizable :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration SimpleDecoration s) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger Tall)) Window

Similar to tallSimpleDecoResizable but with the possibility of setting a custom shrinker and a custom theme.

tallSimpleDwmStyle :: ModifiedLayout (Decoration DwmStyle DefaultShrinker) Tall Window

A Tall layout with the xmonad DwmStyle decoration, default theme and default shrinker.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/tallSimpleDwmStyle.png

tallDwmStyle :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration DwmStyle s) Tall Window

Similar to tallSimpleDwmStyle but with the possibility of setting a custom shrinker and a custom theme.

tallSimpleTabbed :: ModifiedLayout (Decoration TabBarDecoration DefaultShrinker) (ModifiedLayout ResizeScreen Tall) Window

A Tall layout with the xmonad tabbed decoration, default theme and default shrinker.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/tallSimpleTabbed.png

tallTabbed :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration TabBarDecoration s) (ModifiedLayout ResizeScreen Tall) Window

Similar to tallSimpleTabbed but with the possibility of setting a custom shrinker and a custom theme.

Mirror Tall decorated layouts

In this section you will find decorated layouts based on the Mirror layout modifier applied to Tall.

mirrorTallSimpleDefault :: ModifiedLayout (Decoration DefaultDecoration DefaultShrinker) (Mirror Tall) Window

A 'Mirror Tall' layout with the xmonad default decoration, default theme and default shrinker.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/mirrorTallSimpleDefault.png

mirrorTallDefault :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration DefaultDecoration s) (Mirror Tall) Window

Similar to mirrorTallSimpleDefault but with the possibility of setting a custom shrinker and a custom theme.

mirrorTallSimpleDefaultResizable :: ModifiedLayout (Decoration DefaultDecoration DefaultShrinker) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger (Mirror Tall))) Window

A 'Mirror Tall' layout with the xmonad default decoration, default theme and default shrinker, but with the possibility of moving windows with the mouse, and resize/move them with the keyboard.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/mirrorTallSimpleDefaultResizable.png

mirrorTallDefaultResizable :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration DefaultDecoration s) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger (Mirror Tall))) Window

Similar to mirrorTallSimpleDefaultResizable but with the possibility of setting a custom shrinker and a custom theme.

mirrorTallSimpleDeco :: ModifiedLayout (Decoration SimpleDecoration DefaultShrinker) (Mirror Tall) Window

A 'Mirror Tall' layout with the xmonad simple decoration, default theme and default shrinker.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/mirrorTallSimpleDeco.png

mirrorTallDeco :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration SimpleDecoration s) (Mirror Tall) Window

Similar to mirrorTallSimpleDece but with the possibility of setting a custom shrinker and a custom theme.

mirrorTallSimpleDecoResizable :: ModifiedLayout (Decoration SimpleDecoration DefaultShrinker) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger (Mirror Tall))) Window

A 'Mirror Tall' layout with the xmonad simple decoration, default theme and default shrinker, but with the possibility of moving windows with the mouse, and resize/move them with the keyboard.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/mirrorTallSimpleDecoResizable.png

mirrorTallDecoResizable :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration SimpleDecoration s) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger (Mirror Tall))) Window

Similar to mirrorTallSimpleDecoResizable but with the possibility of setting a custom shrinker and a custom theme.

mirrorTallSimpleDwmStyle :: ModifiedLayout (Decoration DwmStyle DefaultShrinker) (Mirror Tall) Window

A 'Mirror Tall' layout with the xmonad DwmStyle decoration, default theme and default shrinker.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/mirrorTallSimpleDwmStyle.png

mirrorTallDwmStyle :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration DwmStyle s) (Mirror Tall) Window

Similar to mirrorTallSimpleDwmStyle but with the possibility of setting a custom shrinker and a custom theme.

mirrorTallSimpleTabbed :: ModifiedLayout (Decoration TabBarDecoration DefaultShrinker) (ModifiedLayout ResizeScreen (Mirror Tall)) Window

A 'Mirror Tall' layout with the xmonad tabbed decoration, default theme and default shrinker.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/mirrorTallSimpleTabbed.png

mirrorTallTabbed :: Shrinker s => s -> Theme -> ModifiedLayout (Decoration TabBarDecoration s) (ModifiedLayout ResizeScreen (Mirror Tall)) Window

Similar to mirrorTallSimpleTabbed but with the possibility of setting a custom shrinker and a custom theme.

Floating decorated layouts

Here you will find decorated layout based on the SimpleFloating layout

floatSimpleSimple :: (Show a, Eq a) => ModifiedLayout (Decoration SimpleDecoration DefaultShrinker) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger SimpleFloat)) a

A simple floating layout where every window is placed according to the window's initial attributes.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/floatSimpleSimple.png

floatSimple :: (Show a, Eq a, Shrinker s) => s -> Theme -> ModifiedLayout (Decoration SimpleDecoration s) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger SimpleFloat)) a

floatSimpleDefault :: (Show a, Eq a) => ModifiedLayout (Decoration DefaultDecoration DefaultShrinker) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger SimpleFloat)) a

This version is decorated with the DefaultDecoration style.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/floatSimpleDefault.png

floatDefault :: (Show a, Eq a, Shrinker s) => s -> Theme -> ModifiedLayout (Decoration DefaultDecoration s) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger SimpleFloat)) a

Same as floatSimpleDefault, but with the possibility of setting a custom shrinker and a custom theme.

floatSimpleDwmStyle :: (Show a, Eq a) => ModifiedLayout (Decoration DwmStyle DefaultShrinker) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger SimpleFloat)) a

This version is decorated with the DwmStyle. Note that this is a keyboard only floating layout.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/floatSimpleDwmStyle.png

floatDwmStyle :: (Show a, Eq a, Shrinker s) => s -> Theme -> ModifiedLayout (Decoration DwmStyle s) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger SimpleFloat)) a

Same as floatSimpleDwmStyle, but with the possibility of setting a custom shrinker and a custom theme.

floatSimpleTabbed :: (Show a, Eq a) => ModifiedLayout (Decoration TabBarDecoration DefaultShrinker) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger SimpleFloat)) a

This version is decorated with the TabbedDecoration style. | Mouse dragging is somehow weird.

Here you can find a screen shot:

http://code.haskell.org/~arossato/xmonadShots/floatSimpleTabbed.png

floatTabbed :: (Show a, Eq a, Shrinker s) => s -> Theme -> ModifiedLayout (Decoration TabBarDecoration s) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger SimpleFloat)) a

Same as floatSimpleTabbed, but with the possibility of setting a custom shrinker and a custom theme.

defaultTheme :: Theme

The default xmonad Theme.

shrinkText :: DefaultShrinker