XMonad.Layout.DragPane

Description

Layouts that splits the screen either horizontally or vertically and shows two windows. The first window is always the master window, and the other is either the currently focused window or the second window in layout order.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Layout.DragPane

Then edit your layoutHook by adding the DragPane layout:

 myLayout = dragPane Horizontal 0.1 0.5 ||| Full ||| etc..
 main = xmonad defaultConfig { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

dragPane :: DragType -> Double -> Double -> DragPane a

data DragPane a

data DragType

Constructors

Horizontal

Vertical