XMonad.Layout.FixedColumn

Description

A layout much like Tall, but using a multiple of a window's minimum resize amount instead of a percentage of screen to decide where to split. This is useful when you usually leave a text editor or terminal in the master pane and like it to be 80 columns wide.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import XMonad.Layout.FixedColumn

Then edit your layoutHook by adding the FixedColumn layout:

myLayout = FixedColumn 1 20 80 10 ||| Full ||| etc..
main = xmonad def { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

data FixedColumn a

A tiling mode based on preserving a nice fixed width window. Supports Shrink, Expand and IncMasterN.

Constructors

FixedColumn !Int !Int !Int !Int