XMonad.Layout.Grid

Description

A simple layout that attempts to put all windows in a square grid.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import XMonad.Layout.Grid

Then edit your layoutHook by adding the Grid layout:

myLayout = Grid ||| Full ||| etc..
main = xmonad def { layoutHook = myLayout }

You can also specify an aspect ratio for Grid to strive for with the GridRatio constructor. For example, if you want Grid to try to make a grid four windows wide and three windows tall, you could use

myLayout = GridRatio (4/3) ||| etc.

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

data Grid a

Constructors

Grid

GridRatio Double

arrange :: Double -> Rectangle -> [a] -> [(a, Rectangle)]

defaultRatio :: Double