XMonad.Layout.IfMax

Description

Provides IfMax layout, which will run one layout if there are maximum N windows on workspace, and another layout, when number of windows is greater than N.

Usage

IfMax layout will run one layout if number of windows on workspace is as maximum N, and else will run another layout.

You can use this module by adding folowing in your xmonad.hs:

import XMonad.Layout.IfMax

Then add layouts to your layoutHook:

myLayoutHook = IfMax 2 Full (Tall ...) ||| ...

In this example, if there are 1 or 2 windows, Full layout will be used; otherwise, Tall layout will be used.

data IfMax l1 l2 w

Constructors

IfMax Int (l1 w) (l2 w)

ifMax

Arguments

:: (LayoutClass l1 w, LayoutClass l2 w)

=> Int
Maximum number of windows for the first layout

-> l1 w
First layout

-> l2 w
Second layout

-> IfMax l1 l2 w

Layout itself