XMonad.Layout.Maximize

Description

Temporarily yanks the focused window out of the layout to mostly fill the screen.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import XMonad.Layout.Maximize

Then edit your layoutHook by adding the Maximize layout modifier:

myLayout = maximize (Tall 1 (3/100) (1/2)) ||| Full ||| etc..)
main = xmonad def { layoutHook = myLayout }

Or, if you want to control the amount of padding placed around the maximized window:

myLayout = maximizeWithPadding 10 (Tall 1 (3/100) (1/2)) ||| Full ||| etc..)
main = xmonad def { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

In the key-bindings, do something like:

       , ((modm, xK_backslash), withFocused (sendMessage . maximizeRestore))
       ...

For detailed instruction on editing the key binding see:

XMonad.Doc.Extending.

maximize :: LayoutClass l Window => l Window -> ModifiedLayout Maximize l Window

maximizeWithPadding :: LayoutClass l Window => Dimension -> l Window -> ModifiedLayout Maximize l Window

Like maximize, but allows you to specify the amount of padding placed around the maximized window.

maximizeRestore :: Window -> MaximizeRestore

data Maximize a

data MaximizeRestore