XMonad.Layout.MosaicAlt

Description

A layout which gives each window a specified amount of screen space relative to the others. Compared to the Mosaic layout, this one divides the space in a more balanced way.

Usage:

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Layout.MosaicAlt
 import qualified Data.Map as M

Then edit your layoutHook by adding the MosaicAlt layout:

 myLayout = MosaicAlt M.empty ||| Full ||| etc..
 main = xmonad defaultConfig { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

In the key-bindings, do something like:

     , ((modm .|. shiftMask  , xK_a    ), withFocused (sendMessage . expandWindowAlt))
     , ((modm .|. shiftMask  , xK_z    ), withFocused (sendMessage . shrinkWindowAlt))
     , ((modm .|. shiftMask  , xK_s    ), withFocused (sendMessage . tallWindowAlt))
     , ((modm .|. shiftMask  , xK_d    ), withFocused (sendMessage . wideWindowAlt))
     , ((modm .|. controlMask, xK_space), sendMessage resetAlt)
     ...

For detailed instruction on editing the key binding see:

XMonad.Doc.Extending.

data MosaicAlt a

Constructors

MosaicAlt Params

shrinkWindowAlt :: Window -> HandleWindowAlt

expandWindowAlt :: Window -> HandleWindowAlt

tallWindowAlt :: Window -> HandleWindowAlt

wideWindowAlt :: Window -> HandleWindowAlt

resetAlt :: HandleWindowAlt

type Params = Map Window Param

data Param

data HandleWindowAlt