XMonad.Layout.MultiDishes

Description

MultiDishes is a layout that stacks groups of extra windows underneath the master windows.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import XMonad.Layout.MultiDishes

Then edit your layoutHook by adding the MultiDishes layout:

myLayout = MultiDishes 2 3 (1/6) ||| Full ||| etc..
main = xmonad def { layoutHook = myLayout }

This is based on the Layout Dishes, but accepts another parameter for the maximum number of dishes allowed within a stack.

MultiDishes x 1 y

is equivalent to > Dishes x y

The stack with the fewest dishes is always on top, so 4 windows with the layout `MultiDishes 1 2 (1/5)` would look like this:

_________
|       |
|   M   |
|_______|
|_______|
|___|___|

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

data MultiDishes a

Constructors

MultiDishes Int Int Rational