XMonad.Layout.Named

Description

A module for assigning a name to a given layout. Deprecated, use XMonad.Layout.Renamed instead.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Layout.Named

Then edit your layoutHook by adding the Named layout modifier to some layout:

 myLayout = named "real big" Full ||| (nameTail $ named "real big" $ Full) ||| etc..
 main = xmonad defaultConfig { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

Note that this module has been deprecated and may be removed in a future release, please use XMonad.Layout.Renamed instead.

named :: String -> l a -> ModifiedLayout Rename l a

(Deprecated) Rename a layout.

nameTail :: l a -> ModifiedLayout Rename l a

(Deprecated) Remove the first word of the name.