XMonad.Layout.PerScreen

Description

Configure layouts based on the width of your screen; use your favorite multi-column layout for wide screens and a full-screen layout for small ones.

Usage

You can use this module by importing it into your ~/.xmonad/xmonad.hs file:

import XMonad.Layout.PerScreen

and modifying your layoutHook as follows (for example):

layoutHook = ifWider 1280 (Tall 1 (3/100) (1/2) ||| Full) Full

Replace any of the layouts with any arbitrarily complicated layout. ifWider can also be used inside other layout combinators.

data PerScreen l1 l2 a

ifWider

Arguments

:: (LayoutClass l1 a, LayoutClass l2 a)

=> Dimension
target screen width

-> l1 a
layout to use when the screen is wide enough

-> l2 a
layout to use otherwise

-> PerScreen l1 l2 a