XMonad.Layout.PositionStoreFloat

Description

A floating layout which has been designed with a dual-head setup in mind. It makes use of XMonad.Util.PositionStore as well as XMonad.Hooks.PositionStoreHooks . Since there is currently no way to move or resize windows with the keyboard alone in this layout, it is adviced to use it in combination with a decoration such as XMonad.Layout.NoFrillsDecoration (to move windows) and the layout modifier XMonad.Layout.BorderResize (to resize windows).

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Layout.PositionStoreFloat
 import XMonad.Layout.NoFrillsDecoration
 import XMonad.Layout.BorderResize

Then edit your layoutHook by adding the PositionStoreFloat layout. Below is a suggestion which uses the mentioned NoFrillsDecoration and BorderResize:

 myLayouts = floatingDeco $ borderResize $ positionStoreFloat ||| etc..
               where floatingDeco l = noFrillsDeco shrinkText defaultTheme l
 main = xmonad defaultConfig { layoutHook = myLayouts }

See the documentation of XMonad.Hooks.PositionStoreHooks on how to add the support hooks.

positionStoreFloat :: PositionStoreFloat a

data PositionStoreFloat a