XMonad.Layout.SimplestFloat

Description

A basic floating layout like SimpleFloat but without the decoration.

Usage:

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import XMonad.Layout.SimplestFloat

Then edit your layoutHook by adding the SimplestFloat layout:

myLayout = simplestFloat ||| Full ||| etc..
main = xmonad def { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

simplestFloat :: Eq a => ModifiedLayout WindowArranger SimplestFloat a

A simple floating layout where every window is placed according to the window's initial attributes.

data SimplestFloat a