XMonad.Layout.Spacing

Description

Add a configurable amount of space around windows.

Usage

You can use this module by importing it into your ~/.xmonad/xmonad.hs file:

 import XMonad.Layout.Spacing

and modifying your layoutHook as follows (for example):

 layoutHook = spacing 2 $ Tall 1 (3/100) (1/2)
                      -- put a 2px space around every window

spacing :: Int -> l a -> ModifiedLayout Spacing l a

Surround all windows by a certain number of pixels of blank space.

data Spacing a

smartSpacing :: Int -> l a -> ModifiedLayout SmartSpacing l a

Surrounds all windows with blank space, except when the window is the only visible window on the current workspace.

data SmartSpacing a