XMonad.Layout.Tabbed

Description

A tabbed layout for the Xmonad Window Manager

Usage:

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import XMonad.Layout.Tabbed

Then edit your layoutHook by adding the Tabbed layout:

myLayout = simpleTabbed ||| Full ||| etc..

or, if you want a specific theme for you tabbed layout:

myLayout = tabbed shrinkText def ||| Full ||| etc..

and then:

main = xmonad def { layoutHook = myLayout }

This layout has hardcoded behaviour for mouse clicks on tab decorations: Left click on the tab switches focus to that window. Middle click on the tab closes the window.

The default Tabbar behaviour is to hide it when only one window is open on the workspace. To have it always shown, use one of the layouts or modifiers ending in Always.

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

You can also edit the default configuration options.

myTabConfig = def { inactiveBorderColor = "#FF0000"
                  , activeTextColor = "#00FF00"}

and

mylayout = tabbed shrinkText myTabConfig ||| Full ||| etc..

simpleTabbed :: ModifiedLayout (Decoration TabbedDecoration DefaultShrinker) Simplest Window

A tabbed layout with the default xmonad Theme.

This is a minimal working configuration:

import XMonad
import XMonad.Layout.Tabbed
main = xmonad def { layoutHook = simpleTabbed }

tabbed :: (Eq a, Shrinker s) => s -> Theme -> ModifiedLayout (Decoration TabbedDecoration s) Simplest a

A layout decorated with tabs and the possibility to set a custom shrinker and theme.

addTabs :: (Eq a, LayoutClass l a, Shrinker s) => s -> Theme -> l a -> ModifiedLayout (Decoration TabbedDecoration s) l a

A layout modifier that uses the provided shrinker and theme to add tabs to any layout.

simpleTabbedAlways :: ModifiedLayout (Decoration TabbedDecoration DefaultShrinker) Simplest Window

tabbedAlways :: (Eq a, Shrinker s) => s -> Theme -> ModifiedLayout (Decoration TabbedDecoration s) Simplest a

addTabsAlways :: (Eq a, LayoutClass l a, Shrinker s) => s -> Theme -> l a -> ModifiedLayout (Decoration TabbedDecoration s) l a

simpleTabbedBottom :: ModifiedLayout (Decoration TabbedDecoration DefaultShrinker) Simplest Window

A bottom-tabbed layout with the default xmonad Theme.

tabbedBottom :: (Eq a, Shrinker s) => s -> Theme -> ModifiedLayout (Decoration TabbedDecoration s) Simplest a

A layout decorated with tabs at the bottom and the possibility to set a custom shrinker and theme.

addTabsBottom :: (Eq a, LayoutClass l a, Shrinker s) => s -> Theme -> l a -> ModifiedLayout (Decoration TabbedDecoration s) l a

A layout modifier that uses the provided shrinker and theme to add tabs to the bottom of any layout.

simpleTabbedLeft :: ModifiedLayout (Decoration TabbedDecoration DefaultShrinker) Simplest Window

A side-tabbed layout with the default xmonad Theme.

tabbedLeft :: (Eq a, Shrinker s) => s -> Theme -> ModifiedLayout (Decoration TabbedDecoration s) Simplest a

A layout decorated with tabs and the possibility to set a custom shrinker and theme.

addTabsLeft :: (Eq a, LayoutClass l a, Shrinker s) => s -> Theme -> l a -> ModifiedLayout (Decoration TabbedDecoration s) l a

A layout modifier that uses the provided shrinker and theme to add tabs to the side of any layout.

simpleTabbedRight :: ModifiedLayout (Decoration TabbedDecoration DefaultShrinker) Simplest Window

A side-tabbed layout with the default xmonad Theme.

tabbedRight :: (Eq a, Shrinker s) => s -> Theme -> ModifiedLayout (Decoration TabbedDecoration s) Simplest a

A layout decorated with tabs and the possibility to set a custom shrinker and theme.

addTabsRight :: (Eq a, LayoutClass l a, Shrinker s) => s -> Theme -> l a -> ModifiedLayout (Decoration TabbedDecoration s) l a

A layout modifier that uses the provided shrinker and theme to add tabs to the side of any layout.

simpleTabbedBottomAlways :: ModifiedLayout (Decoration TabbedDecoration DefaultShrinker) Simplest Window

A bottom-tabbed layout with the default xmonad Theme.

tabbedBottomAlways :: (Eq a, Shrinker s) => s -> Theme -> ModifiedLayout (Decoration TabbedDecoration s) Simplest a

addTabsBottomAlways :: (Eq a, LayoutClass l a, Shrinker s) => s -> Theme -> l a -> ModifiedLayout (Decoration TabbedDecoration s) l a

simpleTabbedLeftAlways :: ModifiedLayout (Decoration TabbedDecoration DefaultShrinker) Simplest Window

A side-tabbed layout with the default xmonad Theme.

tabbedLeftAlways :: (Eq a, Shrinker s) => s -> Theme -> ModifiedLayout (Decoration TabbedDecoration s) Simplest a

A layout decorated with tabs and the possibility to set a custom shrinker and theme.

addTabsLeftAlways :: (Eq a, LayoutClass l a, Shrinker s) => s -> Theme -> l a -> ModifiedLayout (Decoration TabbedDecoration s) l a

simpleTabbedRightAlways :: ModifiedLayout (Decoration TabbedDecoration DefaultShrinker) Simplest Window

A side-tabbed layout with the default xmonad Theme.

tabbedRightAlways :: (Eq a, Shrinker s) => s -> Theme -> ModifiedLayout (Decoration TabbedDecoration s) Simplest a

A layout decorated with tabs and the possibility to set a custom shrinker and theme.

addTabsRightAlways :: (Eq a, LayoutClass l a, Shrinker s) => s -> Theme -> l a -> ModifiedLayout (Decoration TabbedDecoration s) l a

data Theme

A Theme is a record of colors, font etc., to customize a DecorationStyle.

For a collection of Themes see XMonad.Util.Themes

Constructors

Theme

Fields

activeColor :: String

Color of the active window

inactiveColor :: String

Color of the inactive window

urgentColor :: String

Color of the urgent window

activeBorderColor :: String

Color of the border of the active window

inactiveBorderColor :: String

Color of the border of the inactive window

urgentBorderColor :: String

Color of the border of the urgent window

activeTextColor :: String

Color of the text of the active window

inactiveTextColor :: String

Color of the text of the inactive window

urgentTextColor :: String

Color of the text of the urgent window

fontName :: String

Font name

decoWidth :: Dimension

Maximum width of the decorations (if supported by the DecorationStyle)

decoHeight :: Dimension

Height of the decorations

windowTitleAddons :: [(String, Align)]

Extra text to appear in a window's title bar. Refer to for a use XMonad.Layout.ImageButtonDecoration

windowTitleIcons :: [([[Bool]], Placement)]

Extra icons to appear in a window's title bar. Inner [Bool] is a row in a icon bitmap.

def :: Default a => a

defaultTheme :: Theme

Deprecated: Use def (from Data.Default, and re-exported by XMonad.Layout.Decoration) instead.

The default xmonad Theme.

data TabbedDecoration a

Constructors

Tabbed Direction2D TabbarShown

shrinkText :: DefaultShrinker

data CustomShrink

Constructors

CustomShrink

class (Read s, Show s) => Shrinker s where

Minimal complete definition

shrinkIt

Methods

shrinkIt :: s -> String -> [String]

data TabbarShown

data Direction2D

Two-dimensional directions:

Constructors

U
Up

D
Down

R
Right

L
Left