XMonad.Layout.VoidBorders

Description

Modifies a layout to set borders to 0 for all windows in the workspace.

Unlike XMonad.Layout.NoBorders, the voidBorders modifier will not restore the window border if the windows are moved to a different workspace or the layout is changed. There is, however, a companion normalBorders modifier which explicitly restores the border.

This modifier's primary use is to eliminate the "border flash" you get while switching workspaces with the XMonad.Layout.NoBorders modifier.

Usage

You can use this module with the following in your xmonad.hs file:

import XMonad.Layout.VoidBorders

and modify the layouts to call voidBorders on the layouts you want to remove borders from windows, and normalBorders on the layouts you want to keep borders for:

layoutHook = ... ||| voidBorders Full ||| normalBorders Tall ...

For more detailed instructions on editing the layoutHook see the tutorial and XMonad.Doc.Extending.

voidBorders :: l Window -> ModifiedLayout VoidBorders l Window

normalBorders :: l Window -> ModifiedLayout NormalBorders l Window