XMonad.Layout.WindowNavigation

Description

WindowNavigation is an extension to allow easy navigation of a workspace.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import XMonad.Layout.WindowNavigation

Then edit your layoutHook by adding the WindowNavigation layout modifier to some layout:

myLayout = windowNavigation (Tall 1 (3/100) (1/2))  ||| Full ||| etc..
main = xmonad def { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

In keybindings:

   , ((modm,                 xK_Right), sendMessage $ Go R)
   , ((modm,                 xK_Left ), sendMessage $ Go L)
   , ((modm,                 xK_Up   ), sendMessage $ Go U)
   , ((modm,                 xK_Down ), sendMessage $ Go D)
   , ((modm .|. controlMask, xK_Right), sendMessage $ Swap R)
   , ((modm .|. controlMask, xK_Left ), sendMessage $ Swap L)
   , ((modm .|. controlMask, xK_Up   ), sendMessage $ Swap U)
   , ((modm .|. controlMask, xK_Down ), sendMessage $ Swap D)

For detailed instruction on editing the key binding see:

XMonad.Doc.Extending.

windowNavigation :: LayoutClass l a => l a -> ModifiedLayout WindowNavigation l a

configurableNavigation :: LayoutClass l a => WNConfig -> l a -> ModifiedLayout WindowNavigation l a

data Navigate

Constructors

Go Direction2D

Swap Direction2D

Move Direction2D

Apply (Window -> X ()) Direction2D
Apply action with destination window

data Direction2D

Two-dimensional directions:

Constructors

U
Up

D
Down

R
Right

L
Left

data MoveWindowToWindow a

Constructors

MoveWindowToWindow a a

navigateColor :: String -> WNConfig

navigateBrightness :: Double -> WNConfig

noNavigateBorders :: WNConfig

defaultWNConfig :: WNConfig

Deprecated: Use def (from Data.Default, and re-exported by XMonad.Layout.WindowNavigation) instead.

def :: Default a => a

data WNConfig

data WindowNavigation a