XMonad.Layout.WindowSwitcherDecoration

Description

A decoration that allows to switch the position of windows by dragging them onto each other.

Usage:

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Layout.WindowSwitcherDecoration
 import XMonad.Layout.DraggingVisualizer

Then edit your layoutHook by adding the WindowSwitcherDecoration to your layout:

 myL = windowSwitcherDecoration shrinkText defaultTheme (draggingVisualizer $ layoutHook defaultConfig)
 main = xmonad defaultConfig { layoutHook = myL }

There is also a version of the decoration that contains buttons like XMonad.Layout.ButtonDecoration. To use that version, you will need to import XMonad.Layout.DecorationAddons as well and modify your layoutHook in the following way:

 import XMonad.Layout.DecorationAddons

 myL = windowSwitcherDecorationWithButtons shrinkText defaultThemeWithButtons (draggingVisualizer $ layoutHook defaultConfig)
 main = xmonad defaultConfig { layoutHook = myL }

Additionaly, there is a version of the decoration that contains image buttons like XMonad.Layout.ImageButtonDecoration. To use that version, you will need to import XMonad.Layout.ImageButtonDecoration as well and modify your layoutHook in the following way:

 import XMonad.Layout.ImageButtonDecoration

 myL = windowSwitcherDecorationWithImageButtons shrinkText defaultThemeWithImageButtons (draggingVisualizer $ layoutHook defaultConfig)
 main = xmonad defaultConfig { layoutHook = myL }

windowSwitcherDecoration :: (Eq a, Shrinker s) => s -> Theme -> l a -> ModifiedLayout (Decoration WindowSwitcherDecoration s) l a

windowSwitcherDecorationWithButtons :: (Eq a, Shrinker s) => s -> Theme -> l a -> ModifiedLayout (Decoration WindowSwitcherDecoration s) l a

windowSwitcherDecorationWithImageButtons :: (Eq a, Shrinker s) => s -> Theme -> l a -> ModifiedLayout (Decoration ImageWindowSwitcherDecoration s) l a

data WindowSwitcherDecoration a

data ImageWindowSwitcherDecoration a