XMonad.Prompt.RunOrRaise

Description

A prompt for XMonad which will run a program, open a file, or raise an already running program, depending on context.

Usage

In your ~/.xmonad/xmonad.hs:

 import XMonad.Prompt
 import XMonad.Prompt.RunOrRaise

In your keybindings add something like:

   , ((modm .|. controlMask, xK_x), runOrRaisePrompt defaultXPConfig)

For detailed instruction on editing the key binding see XMonad.Doc.Extending.

runOrRaisePrompt :: XPConfig -> X ()

data RunOrRaisePrompt