XMonad.Prompt.Ssh

Description

A ssh prompt for XMonad

Usage

In your ~/.xmonad/xmonad.hs:

 import XMonad.Prompt
 import XMonad.Prompt.Ssh

In your keybindings add something like:

   , ((modm .|. controlMask, xK_s), sshPrompt defaultXPConfig)

Keep in mind, that if you want to use the completion you have to disable the HashKnownHosts option in your ssh_config

For detailed instruction on editing the key binding see XMonad.Doc.Extending.

sshPrompt :: XPConfig -> X ()

data Ssh