XMonad.Prompt.Unicode

Description

A prompt for searching unicode characters by name and inserting them into the clipboard.

Requires the file /usr/share/unicode/UnicodeData.txt (shipped in the package unicode-data on Debian) and the xsel tool.

Usage

You can use this module by importing it, along with XMonad.Prompt, into your ~/.xmonad/xmonad.hs file:

import XMonad.Prompt
import XMonad.Prompt.Unicode

and adding an appropriate keybinding, for example:

 , ((modm .|. controlMask, xK_u), unicodePrompt def)

unicodePrompt :: XPConfig -> X ()

Prompt the user for a unicode character to be inserted into the paste buffer of the X server.