XMonad.Prompt.Window

Description

xprompt operations to bring windows to you, and bring you to windows.

Usage

WindowPrompt brings windows to you and you to windows. That is to say, it pops up a prompt with window names, in case you forgot where you left your XChat. It also offers helpers to build the subset of windows which is used for the prompt completion.

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import XMonad.Prompt
import XMonad.Prompt.Window

and in the keys definition:

, ((modm .|. shiftMask, xK_g     ), windowPrompt def Goto wsWindows)
, ((modm .|. shiftMask, xK_b     ), windowPrompt def Bring allWindows)

The autoComplete option is a handy complement here:

, ((modm .|. shiftMask, xK_g     ), windowPrompt
                                       def { autoComplete = Just 500000 }
                                       Goto allWindows)

The '500000' is the number of microseconds to pause before sending you to your new window. This is useful so that you don't accidentally send some keystrokes to the selected client.

For detailed instruction on editing the key binding see XMonad.Doc.Extending.

data WindowPrompt

Constructors

Goto

Bring

BringCopy

BringToMaster

windowPrompt :: XPConfig -> WindowPrompt -> XWindowMap -> X ()

Pops open a prompt with window titles belonging to winmap. Choose one, and an action is applied on the selected window, according to WindowPrompt.

allWindows :: XWindowMap

A helper to get the map of all windows.

wsWindows :: XWindowMap

A helper to get the map of windows of the current workspace.

type XWindowMap = X (Map String Window)

A Map where keys are pretty printable window names and values are Xmonad windows identifier.

Deprecated

windowPromptGoto :: XPConfig -> X ()

Deprecated. Use windowPrompt instead.

windowPromptBring :: XPConfig -> X ()

Deprecated. Use windowPrompt instead.

windowPromptBringCopy :: XPConfig -> X ()

Deprecated. Use windowPrompt instead.