XMonad.Prompt.XMonad

Description

A prompt for running XMonad commands

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Prompt
 import XMonad.Prompt.XMonad

in your keybindings add:

   , ((modm .|. controlMask, xK_x), xmonadPrompt defaultXPConfig)

For detailed instruction on editing the key binding see XMonad.Doc.Extending.

xmonadPrompt :: XPConfig -> X ()

xmonadPromptC :: [(String, X ())] -> XPConfig -> X ()

An xmonad prompt with a custom command list

data XMonad