XMonad.Prompt.Zsh

Description

A version of XMonad.Prompt.Shell that lets you access the awesome power of Zsh completions in your xmonad prompt

Usage

Grab the capture.zsh script to capture zsh completions from https://github.com/Valodim/zsh-capture-completion

In your xmonad.hs:

import XMonad.Prompt
import XMonad.Prompt.Zsh

In your keybindings add something like:

  , ((modm .|. controlMask, xK_x), zshPrompt def "/path/to/capture.zsh")

For detailed instruction on editing the key binding see the tutorial.

data Zsh

Constructors

Zsh

zshPrompt :: XPConfig -> FilePath -> X ()

Utility functions

getZshCompl :: FilePath -> String -> IO [String]

stripZsh :: String -> String

Removes the argument description from the zsh completion