XMonad.Util.DynamicScratchpads

Description

Dynamically declare any window as a scratchpad.

Usage

Allows you to dynamically declare windows as scratchpads. You can bind a key to make a window start/stop being a scratchpad, and another key to spawn/hide that scratchpad.

Like with XMonad.Util.NamedScratchpad, you have to have a workspace called NSP, where hidden scratchpads will be moved to.

You can declare dynamic scrachpads in your xmonad.hs like so:

import XMonad.Util.DynamicScratchpads

, ((modm .|. shiftMask, xK_a), withFocused $ makeDynamicSP "dyn1") , ((modm .|. shiftMask, xK_b), withFocused $ makeDynamicSP "dyn2") , ((modm , xK_a), spawnDynamicSP "dyn1") , ((modm , xK_b), spawnDynamicSP "dyn2")

makeDynamicSP

Arguments

:: String
Scratchpad name

-> Window
Window to be made a scratchpad

-> X ()

Makes a window a dynamic scratchpad with the given name, or stop a window | from being a dynamic scratchpad, if it already is.

spawnDynamicSP

Arguments

:: String
Scratchpad name

-> X ()

Spawn the specified dynamic scratchpad