XMonad.Util.Font

Description

A module for abstracting a font facility over Core fonts and Xft

Usage:

See XMonad.Layout.Tabbed or XMonad.Prompt for usage examples

data XMonadFont

Constructors

Core FontStruct

Utf8 FontSet

Xft XftFont

initXMF :: String -> X XMonadFont

When initXMF gets a font name that starts with 'xft:' it switches to the Xft backend Example: 'xft: Sans-10'

releaseXMF :: XMonadFont -> X ()

initCoreFont :: String -> X FontStruct

Given a fontname returns the font structure. If the font name is not valid the default font will be loaded and returned.

releaseCoreFont :: FontStruct -> X ()

initUtf8Font :: String -> X FontSet

releaseUtf8Font :: FontSet -> X ()

data Align

String position

Constructors

AlignCenter

AlignRight

AlignLeft

AlignRightOffset Int

stringPosition :: (Functor m, MonadIO m) => Display -> XMonadFont -> Rectangle -> Align -> String -> m (Position, Position)

Return the string x and y Position in a Rectangle, given a FontStruct and the Alignment

textWidthXMF :: MonadIO m => Display -> XMonadFont -> String -> m Int

textExtentsXMF :: MonadIO m => XMonadFont -> String -> m (Int32, Int32)

printStringXMF :: (Functor m, MonadIO m) => Display -> Drawable -> XMonadFont -> GC -> String -> String -> Position -> Position -> String -> m ()

stringToPixel :: (Functor m, MonadIO m) => Display -> String -> m Pixel

Get the Pixel value for a named color: if an invalid name is given the black pixel will be returned.

pixelToString :: MonadIO m => Display -> Pixel -> m String

Convert a Pixel into a String.

fi :: (Integral a, Num b) => a -> b

Short-hand for fromIntegral