XMonad.Util.Grab

Description

This module should not be directly used by users. Its purpose is to facilitate grabbing and ungrabbing keys.

Usage

This module should not be directly used by users. Its purpose is to facilitate grabbing and ungrabbing keys.

grabKP :: KeyMask -> KeyCode -> X ()

A more convenient version of grabKey.

ungrabKP :: KeyMask -> KeyCode -> X ()

A more convenient version of ungrabKey.

grabUngrab

Arguments

:: [(KeyMask, KeySym)]
Keys to grab

-> [(KeyMask, KeySym)]
Keys to ungrab

-> X ()

A convenience function to grab and ungrab keys

grab :: [(KeyMask, KeySym)] -> X ()

A convenience function to grab keys. This also ungrabs all previously grabbed keys.

customRegrabEvHook :: X () -> Event -> X All

An event hook that runs a custom action to regrab the necessary keys.