XMonad.Util.SpawnOnce

Description

A module for spawning a command once, and only once. Useful to start status bars and make session settings inside startupHook.

Documentation

spawnOnce :: String -> X ()

The first time spawnOnce is executed on a particular command, that command is executed. Subsequent invocations for a command do nothing.