XMonad.Util.Themes

Description

A (hopefully) growing collection of themes for decorated layouts.

Usage

This module stores some user contributed themes which can be used with decorated layouts (such as Tabbed). (Note that these themes only apply to decorated layouts, such as those found in XMonad.Layout.Tabbed and XMonad.Layout.DecorationMadness; they do not apply to xmonad as a whole.)

If you want to use one of them with one of your decorated layouts, you need to substitute defaultTheme with, for instance, (theme smallClean).

Here is an example:

 import XMonad
 import XMonad.Util.Themes
 import XMonad.Layout.Tabbed

 myLayout = tabbed shrinkText (theme smallClean)

 main = xmonad defaultConfig {layoutHook = myLayout}

If you have a theme you would like to share, adding it to this module is very easy.

You can use xmonadTheme or smallClean as a template.

At the present time only the themeName field is used. But please provide all the other information, which will be used at a later time.

Please, remember to add your theme to the list of exported functions, and to the listOfThemes.

Thanks for your contribution!

listOfThemes :: [ThemeInfo]

ppThemeInfo :: ThemeInfo -> String

xmonadTheme :: ThemeInfo

The default xmonad theme, by David Roundy.

smallClean :: ThemeInfo

Small decorations with a Ion3 remembrance, by Andrea Rossato.

robertTheme :: ThemeInfo

Ffrom Robert Manea's prompt theme.

deiflTheme :: ThemeInfo

deifl's Theme, by deifl.

oxymor00nTheme :: ThemeInfo

oxymor00n's theme, by Tom Rauchenwald.

donaldTheme :: ThemeInfo

Don's preferred colors - from DynamicLog...;)

wfarrTheme :: ThemeInfo

kavonForestTheme :: ThemeInfo

Forest colours, by Kathryn Andersen

kavonLakeTheme :: ThemeInfo

Lake (blue/green) colours, by Kathryn Andersen

kavonPeacockTheme :: ThemeInfo

Peacock colours, by Kathryn Andersen

kavonVioGreenTheme :: ThemeInfo

Violet-Green colours, by Kathryn Andersen

kavonBluesTheme :: ThemeInfo

Blue colours, by Kathryn Andersen

kavonAutumnTheme :: ThemeInfo

Autumn colours, by Kathryn Andersen

kavonFireTheme :: ThemeInfo

Fire colours, by Kathryn Andersen

kavonChristmasTheme :: ThemeInfo

Christmas colours, by Kathryn Andersen

data ThemeInfo

Constructors

TI

Fields

themeName :: String

themeAuthor :: String

themeDescription :: String

theme :: Theme