XMonad.Util.WindowProperties

Description

EDSL for specifying window properties; various utilities related to window properties.

EDSL for window properties

Allows to specify window properties, such as title, classname or resource, and to check them.

In contrast to ManageHook properties, these are instances of Show and Read, so they can be used in layout definitions etc. For example usage see XMonad.Layout.IM

data Property

Most of the property constructors are quite self-explaining.

Constructors

Title String

ClassName String

Resource String

Role String
WM_WINDOW_ROLE property

Machine String
WM_CLIENT_MACHINE property

And Property Property

Or Property Property

Not Property

Const Bool

hasProperty :: Property -> Window -> X Bool

Does given window have this property?

focusedHasProperty :: Property -> X Bool

Does the focused window have this property?

allWithProperty :: Property -> X [Window]

Find all existing windows with specified property

propertyToQuery :: Property -> Query Bool

Convert property to Query Bool (see XMonad.ManageHook)

Helper functions

getProp32 :: Atom -> Window -> X (Maybe [CLong])

Get a window property from atom

getProp32s :: String -> Window -> X (Maybe [CLong])

Get a window property from string