XMonad.Actions.CycleRecentWS

Description

Provides bindings to cycle through most recently used workspaces with repeated presses of a single key (as long as modifier key is held down). This is similar to how many window managers handle window switching.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs file:

import XMonad.Actions.CycleRecentWS

  , ((modm, xK_Tab), cycleRecentWS [xK_Alt_L] xK_Tab xK_grave)

For detailed instructions on editing your key bindings, see XMonad.Doc.Extending.

cycleRecentWS

Arguments

:: [KeySym]
A list of modifier keys used when invoking this action. As soon as one of them is released, the final switch is made.

-> KeySym
Key used to switch to next (less recent) workspace.

-> KeySym
Key used to switch to previous (more recent) workspace. If it's the same as the nextWorkspace key, it is effectively ignored.

-> X ()

Cycle through most recent workspaces with repeated presses of a key, while a modifier key is held down. The recency of workspaces previewed while browsing to the target workspace is not affected. That way a stack of most recently used workspaces is maintained, similarly to how many window managers handle window switching. For best effects use the same modkey+key combination as the one used to invoke this action.

cycleWindowSets

Arguments

:: (WindowSet -> [WindowSet])
A function used to create a list of WindowSets to choose from

-> [KeySym]
A list of modifier keys used when invoking this action. As soon as one of them is released, the final WindowSet is chosen and the action exits.

-> KeySym
Key used to preview next WindowSet from the list of generated options

-> KeySym
Key used to preview previous WindowSet from the list of generated options. If it's the same as nextOption key, it is effectively ignored.

-> X ()

Cycle through a finite list of WindowSets with repeated presses of a key, while a modifier key is held down. For best effects use the same modkey+key combination as the one used to invoke this action.