XMonad.Actions.CycleSelectedLayouts

Description

This module allows to cycle through the given subset of layouts.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import XMonad
import XMonad.Actions.CycleSelectedLayouts

  , ((modm,  xK_t ),   cycleThroughLayouts ["Tall", "Mirror Tall"])

cycleThroughLayouts :: [String] -> X ()

If the current layout is in the list, cycle to the next layout. Otherwise, apply the first layout from list.