XMonad.Actions.FindEmptyWorkspace

Description

Find an empty workspace.

Usage

To use, import this module into your ~/.xmonad/xmonad.hs:

   import XMonad.Actions.FindEmptyWorkspace

and add the desired keybindings, for example:

   , ((modm,                xK_m    ), viewEmptyWorkspace)
   , ((modm .|. shiftMask,  xK_m    ), tagToEmptyWorkspace)

Now you can jump to an empty workspace with mod-m. Mod-shift-m will tag the current window to an empty workspace and view it.

For detailed instructions on editing your key bindings, see XMonad.Doc.Extending.

viewEmptyWorkspace :: X ()

Find and view an empty workspace. Do nothing if all workspaces are in use.

tagToEmptyWorkspace :: X ()

Tag current window to an empty workspace and view it. Do nothing if all workspaces are in use.

sendToEmptyWorkspace :: X ()

Send current window to an empty workspace. Do nothing if all workspaces are in use.