XMonad.Actions.FocusNth

Description

Focus the nth window of the current workspace.

Usage

Add the import to your xmonad.hs:

import XMonad.Actions.FocusNth

Then add appropriate keybindings, for example:

-- mod4-[1..9] @@ Switch to window N
++ [((modm, k), focusNth i)
    | (i, k) <- zip [0 .. 8] [xK_1 ..]]

For detailed instructions on editing your key bindings, see the tutorial.

focusNth :: Int -> X ()

Give focus to the nth window of the current workspace.

focusNth' :: Int -> Stack a -> Stack a

swapNth :: Int -> X ()

Swap current window with nth. Focus stays in the same position

swapNth' :: Int -> Stack a -> Stack a