XMonad.Actions.MouseResize

Description

A layout modifier to resize windows with the mouse by grabbing the window's lower right corner.

This module must be used together with XMonad.Layout.WindowArranger.

Usage:

Usually this module is used to create layouts, but you can also use it to resize windows in any layout, together with the XMonad.Layout.WindowArranger. For usage example see XMonad.Layout.SimpleFloat or XMonad.Layout.DecorationMadness.

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Actions.MouseResize
 import XMonad.Layout.WindowArranger

Then edit your layoutHook by modifying a given layout:

 myLayout = mouseResize $ windowArrange $ layoutHook defaultConfig

and then:

 main = xmonad defaultConfig { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

mouseResize :: l a -> ModifiedLayout MouseResize l a

data MouseResize a

Constructors

MR [((a, Rectangle), Maybe a)]