XMonad.Actions.PerWorkspaceKeys

Description

Define key-bindings on per-workspace basis.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Actions.PerWorkspaceKeys

  ,((0, xK_F2), bindOn [("1", spawn "rxvt"), ("2", spawn "xeyes"), ("", spawn "xmessage hello")])

For detailed instructions on editing your key bindings, see XMonad.Doc.Extending.

chooseAction :: (String -> X ()) -> X ()

Uses supplied function to decide which action to run depending on current workspace name.

bindOn :: [(String, X ())] -> X ()

If current workspace is listed, run appropriate action (only the first match counts!) If it isn't listed, then run default action (marked with empty string, ""), or do nothing if default isn't supplied.