XMonad.Actions.Promote

Description

Alternate promote function for xmonad.

Moves the focused window to the master pane. All other windows retain their order. If focus is in the master, swap it with the next window in the stack. Focus stays in the master.

Usage

You can use this module with the following in your xmonad.hs:

   import XMonad.Actions.Promote

then add a keybinding or substitute promote in place of swapMaster:

  , ((modm,               xK_Return), promote)

For detailed instructions on editing your key bindings, see the tutorial.

promote :: X ()

Move the focused window to the master pane. All other windows retain their order. If focus is in the master, swap it with the next windo in the stack. Focus stays in the master.