XMonad.Actions.RandomBackground

Description

An action to start terminals with a random background color

Usage

Add to your keybindings something like:

,((modm .|. shiftMask, xK_Return), randomBg $ HSV 0xff 0x20

randomBg' :: MonadIO m => RandomColor -> m String

randomBg' produces a random hex number in the form '#xxyyzz'

randomBg :: RandomColor -> X ()

randomBg starts a terminal with the background color taken from randomBg'

This depends on the your terminal configuration field accepting an argument like -bg '#ff0023'

data RandomColor

RandomColor fixes constraints when generating random colors. All parameters should be in the range 0 -- 0xff

Constructors

RGB Int Int
specify the minimum and maximum lowest values for each color channel.

HSV Double Double
specify the saturation and value, leaving the hue random.