XMonad.Actions.RotSlaves

Description

Rotate all windows except the master window and keep the focus in place.

Documentation

To use this module, import it with:

import XMonad.Actions.RotSlaves

and add whatever keybindings you would like, for example:

, ((modm .|. shiftMask, xK_Tab   ), rotSlavesUp)

This operation will rotate all windows except the master window, while the focus stays where it is. It is useful together with the TwoPane layout (see XMonad.Layout.TwoPane).

For detailed instructions on editing your key bindings, see XMonad.Doc.Extending.

rotSlaves' :: ([a] -> [a]) -> Stack a -> Stack a

The actual rotation, as a pure function on the window stack.

rotSlavesUp :: X ()

Rotate the windows in the current stack, excluding the first one (master).

rotSlavesDown :: X ()

Rotate the windows in the current stack, excluding the first one (master).

rotAll' :: ([a] -> [a]) -> Stack a -> Stack a

The actual rotation, as a pure function on the window stack.

rotAllUp :: X ()

Rotate all the windows in the current stack.

rotAllDown :: X ()

Rotate all the windows in the current stack.