XMonad.Actions.WindowBringer

Description

dmenu operations to bring windows to you, and bring you to windows. That is to say, it pops up a dmenu with window names, in case you forgot where you left your XChat.

Usage

Import the module into your ~/.xmonad/xmonad.hs:

 import XMonad.Actions.WindowBringer

and define appropriate key bindings:

 , ((modm .|. shiftMask, xK_g     ), gotoMenu)
 , ((modm .|. shiftMask, xK_b     ), bringMenu)

For detailed instructions on editing your key bindings, see XMonad.Doc.Extending.

gotoMenu :: X ()

Pops open a dmenu with window titles. Choose one, and you will be taken to the corresponding workspace.

gotoMenu' :: String -> X ()

Pops open an application with window titles given over stdin. Choose one, and you will be taken to the corresponding workspace.

gotoMenuArgs :: [String] -> X ()

Pops open a dmenu with window titles. Choose one, and you will be taken to the corresponding workspace. This version takes a list of arguments to pass to dmenu.

gotoMenuArgs' :: String -> [String] -> X ()

Pops open an application with window titles given over stdin. Choose one, and you will be taken to the corresponding workspace. This version takes a list of arguments to pass to dmenu.

bringMenu :: X ()

Pops open a dmenu with window titles. Choose one, and it will be dragged, kicking and screaming, into your current workspace.

bringMenu' :: String -> X ()

Pops open an application with window titles given over stdin. Choose one, and it will be dragged, kicking and screaming, into your current workspace.

bringMenuArgs :: [String] -> X ()

Pops open a dmenu with window titles. Choose one, and it will be dragged, kicking and screaming, into your current workspace. This version takes a list of arguments to pass to dmenu.

bringMenuArgs' :: String -> [String] -> X ()

Pops open an application with window titles given over stdin. Choose one, and it will be dragged, kicking and screaming, into your current workspace. This version allows arguments to the chooser to be specified.

windowMap :: X (Map String Window)

A map from window names to Windows.

bringWindow :: Window -> WindowSet -> WindowSet

Brings the specified window into the current workspace.