XMonad.Actions.WindowNavigation

Description

This is a rewrite of XMonad.Layout.WindowNavigation. WindowNavigation lets you assign keys to move up/down/left/right, based on actual cartesian window coordinates, rather than just going j/k on the stack.

This module is experimental. You'll have better luck with the original.

This module differs from the other in a few ways:

You can go up/down/left/right across multiple screens.

It doesn't provide little border colors for your neighboring windows.

It doesn't provide the 'Move' action, which seems to be related to the XMonad.Layout.Combo extension.

It tries to be slightly smarter about tracking your current position.

Configuration is different.

Usage

To use it, you're going to apply the withWindowNavigation function. withWindowNavigation performs some IO operations, so the syntax you'll use is the same as the spawnPipe example in XMonad.Hooks.DynamicLog. In particular:

main = do
    config <- withWindowNavigation (xK_w, xK_a, xK_s, xK_d)
            $ def { ... }
    xmonad config

Or, for the brave souls:

main = xmonad =<< withWindowNavigation (xK_w, xK_a, xK_s, xK_d)
            $ def { ... }

Here, we pass in the keys for navigation in counter-clockwise order from up. It creates keybindings for modMask to move to window, and modMask .|. shiftMask to swap windows.

If you want more flexibility over your keybindings, you can use withWindowNavigationKeys, which takes a list of keys-esque entries rather than a tuple of the four directional keys. See the source code of withWindowNavigation for an example.

withWindowNavigation :: (KeySym, KeySym, KeySym, KeySym) -> XConfig l -> IO (XConfig l)

withWindowNavigationKeys :: [((KeyMask, KeySym), WNAction)] -> XConfig l -> IO (XConfig l)

data WNAction

Constructors

WNGo Direction2D

WNSwap Direction2D

go :: IORef WNState -> Direction2D -> X ()

swap :: IORef WNState -> Direction2D -> X ()

data Direction2D

Two-dimensional directions:

Constructors

U
Up

D
Down

R
Right

L
Left

type WNState = Map WorkspaceId Point