XMonad.Config.Arossato

Description

This module specifies my xmonad defaults.

Usage

The simplest way to use this configuration module is to use an ~/.xmonad/xmonad.hs like this:

module Main (main) where

import XMonad
import XMonad.Config.Arossato (arossatoConfig)

main :: IO ()
main = xmonad =<< arossatoConfig

NOTE: that I'm using xmobar and, if you don't have xmobar in your PATH, this configuration will produce an error and xmonad will not start. If you don't want to install xmobar get rid of this line at the beginning of arossatoConfig.

You can use this module also as a starting point for writing your own configuration module from scratch. Save it as your ~/.xmonad/xmonad.hs and:

Change the module name from

module XMonad.Config.Arossato
    ( -- * Usage
      -- $usage
      arossatoConfig
    ) where

to

module Main where

Add a line like:

main = xmonad =<< arossatoConfig

Start playing with the configuration options...;)

arossatoConfig :: MonadIO m => m (XConfig (ModifiedLayout AvoidStruts (Choose (ModifiedLayout (Decoration SimpleDecoration DefaultShrinker) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger SimpleFloat))) (Choose (ModifiedLayout WithBorder (ModifiedLayout (Decoration TabbedDecoration DefaultShrinker) Simplest)) (ModifiedLayout WindowArranger (Choose (ModifiedLayout Magnifier Tall) (Choose (ModifiedLayout WithBorder Full) (Choose (Mirror Tall) Accordion))))))))