XMonad.Config.Gnome

Description

This module provides a config suitable for use with the GNOME desktop environment.

Usage

To use this module, start with the following xmonad.hs:

import XMonad
import XMonad.Config.Gnome

main = xmonad gnomeConfig

For examples of how to further customize gnomeConfig see XMonad.Config.Desktop.

gnomeConfig :: XConfig (ModifiedLayout AvoidStruts (Choose Tall (Choose (Mirror Tall) Full)))

gnomeRun :: X ()

Launch the "Run Application" dialog. gnome-panel must be running for this to work.

gnomeRegister :: MonadIO m => m ()

Register xmonad with gnome. 'dbus-send' must be in the $PATH with which xmonad is started.

This action reduces a delay on startup only only if you have configured gnome-session>=2.26: to start xmonad with a command as such:

gconftool-2 -s /desktop/gnome/session/required_components/windowmanager xmonad --type string

desktopLayoutModifiers :: LayoutClass l a => l a -> ModifiedLayout AvoidStruts l a