XMonad.Config.Mate

Description

This module provides a config suitable for use with the MATE desktop environment.

Usage

To use this module, start with the following ~/.xmonad/xmonad.hs:

import XMonad
import XMonad.Config.Mate

main = xmonad mateConfig

For examples of how to further customize mateConfig see XMonad.Config.Desktop.

mateConfig :: XConfig (ModifiedLayout AvoidStruts (Choose Tall (Choose (Mirror Tall) Full)))

mateRun :: X ()

Launch the "Run Application" dialog. mate-panel must be running for this to work. partial application for existing keybinding compatibility.

matePanel :: String -> X ()

Launch a panel action. Either the "Run Application" dialog ("run_dialog" parameter, see above) or the main menu ("main_menu" parameter). mate-panel must be running for this to work.

mateRegister :: MonadIO m => m ()

Register xmonad with mate. 'dbus-send' must be in the $PATH with which xmonad is started.

This action reduces a delay on startup only if you have configured mate-session to start xmonad with a command such as (check local documentation):

dconf write /org/mate/desktop/session/required_components/windowmanager "'xmonad'"

(the extra quotes are required by dconf)

mateLogout :: MonadIO m => m ()

Display MATE logout dialog. This is the default mod-q action.

mateShutdown :: MonadIO m => m ()

Display MATE shutdown dialog. You can override mod-q to invoke this, or bind it to another key if you prefer.

desktopLayoutModifiers :: LayoutClass l a => l a -> ModifiedLayout AvoidStruts l a