XMonad.Hooks.DebugEvents

Description

Module to dump diagnostic information about X11 events received by xmonad. This is incomplete due to Event being incomplete and not providing information about a number of events, and enforcing artificial constraints on others (for example ClientMessage); the X11 package will require a number of changes to fix these problems.

Documentation

debugEventsHook :: Event -> X All

Event hook to dump all received events. You should probably not use this unconditionally; it will produce massive amounts of output.