XMonad.Hooks.DebugKeyEvents

Description

A debugging module to track key events, useful when you can't tell whether xmonad is processing some or all key events.

Usage

Add this to your handleEventHook to print received key events to the log (the console if you use startx/xinit, otherwise usually ~/.xsession-errors).

     , handleEventHook = debugKeyEvents

If you already have a handleEventHook then you should append it:

     , handleEventHook = ... <> debugKeyEvents

Logged key events look like:

keycode 53 sym 120 (0x78, "x") mask 0x0 () clean 0x0 ()

The mask and clean indicate the modifiers pressed along with the key; mask is raw, and clean is what xmonad sees after sanitizing it (removing numberLockMask, etc.)

For more detailed instructions on editing the logHook see:

XMonad.Doc.Extending

debugKeyEvents :: Event -> X All

Print key events to stderr for debugging