XMonad.Hooks.DebugStack

Description

Dump the state of the StackSet. A logHook and handleEventHook are also provided.

Documentation

debugStack :: X ()

Print the state of the current window stack to stderr, which for most installations goes to ~/.xsession-errors. XMonad.Util.DebugWindow is used to display the individual windows.

debugStackString :: X String

Dump the state of the current StackSet as a multiline String. stack [ mm ,(*) ww , ww ] float { ww , ww }

One thing I'm not sure of is where the zipper is when focus is on a floating window.

debugStackLogHook :: X ()

The above packaged as a logHook. (Currently this is identical.)

debugStackEventHook :: Event -> X All

The above packaged as a handleEventHook. You almost certainly do not want to use this unconditionally, as it will cause massive amounts of output and possibly slow xmonad down severely.