XMonad.Hooks.Script

Description

Provides a simple interface for running a ~/.xmonad/hooks script with the name of a hook.

Usage

This module allows you to run a centrally located script with the text name of a hook. The script is assumed to be located at ~/.xmonad/hooks.

For example, if you wanted to run the hook startup in your script every time your startup hook ran, you could modify your xmonad config as such:

 main = xmonad $ defaultConfig {
   ...
   startupHook = execScriptHook "startup"
   ...
   }

Now, every time the startup hook runs, the command ~/.xmonad/hooks startup will also.

Script Hook Interface

execScriptHook :: MonadIO m => String -> m ()

Execute a named script hook