XMonad.Layout.AutoMaster

Description

Provides layout modifier AutoMaster. It separates screen in two parts - master and slave. Size of slave area automatically changes depending on number of slave windows.

Usage

This module defines layout modifier named autoMaster. It separates screen in two parts - master and slave. Master windows are arranged in one row, in slave area underlying layout is run. Size of slave area automatically increases when number of slave windows is increasing.

You can use this module by adding folowing in your xmonad.hs:

 import XMonad.Layout.AutoMaster

Then add layouts to your layoutHook:

 myLayoutHook = autoMaster 1 (1/100) Grid ||| ...

In this example, master area by default contains 1 window (you can change this number in runtime with usual IncMasterN message), changing slave area size with 1100 on each ShrinkExpand message.

autoMaster :: LayoutClass l a => Int -> Float -> l a -> ModifiedLayout AutoMaster l a

User interface function

data AutoMaster a

Data type for layout modifier