XMonad.Layout.BoringWindows

Description

BoringWindows is an extension to allow windows to be marked boring

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Layout.BoringWindows

Then edit your layoutHook by adding the layout modifier:

 myLayout = boringWindows (Full ||| etc..)
 main = xmonad defaultConfig { layoutHook = myLayout }

Then to your keybindings, add:

 , ((modm, xK_j), focusUp)
 , ((modm, xK_k), focusDown)
 , ((modm, xK_m), focusMaster)

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

boringWindows :: (LayoutClass l a, Eq a) => l a -> ModifiedLayout BoringWindows l a

boringAuto :: (LayoutClass l a, Eq a) => l a -> ModifiedLayout BoringWindows l a

Mark windows that are not given rectangles as boring

markBoring :: X ()

clearBoring :: X ()

focusUp :: X ()

focusDown :: X ()

focusMaster :: X ()

data UpdateBoring

UpdateBoring is sent before attempting to view another boring window, so that layouts have a chance to mark boring windows.

Constructors

UpdateBoring

data BoringMessage

Constructors

Replace String [Window]

Merge String [Window]

data BoringWindows a

Tips

variant of Full

An alternative to Full is XMonad.Layout.Simplest. Less windows are ignored by focusUp and focusDown. This may be helpful when you want windows to be uninteresting by some other layout modifier (ex. XMonad.Layout.Minimize)