XMonad.Layout.CenteredMaster

Description

Two layout modifiers. centerMaster places master window at center, on top of all other windows, which are managed by base layout. topRightMaster is similar, but places master window in top right corner instead of center.

Usage

This module defines two new layout modifiers: centerMaster and topRightMaster. centerMaster places master window at center of screen, on top of others. All other windows in background are managed by base layout. topRightMaster is like centerMaster, but places master window in top right corner instead of center.

Yo can use this module by adding folowing in your xmonad.hs:

import XMonad.Layout.CenteredMaster

Then add layouts to your layoutHook:

myLayoutHook = centerMaster Grid ||| ...

centerMaster :: LayoutClass l a => l a -> ModifiedLayout CenteredMaster l a

Modifier that puts master window in center, other windows in background are managed by given layout

topRightMaster :: LayoutClass l a => l a -> ModifiedLayout TopRightMaster l a

Modifier that puts master window in top right corner, other windows in background are managed by given layout

data CenteredMaster a

Data type for LayoutModifier

data TopRightMaster a