XMonad.Layout.Circle

Description

Deprecated: Use XMonad.Layout.CircleEx instead

Circle is an elliptical, overlapping layout, by Peter De Wachter

Usage

You can use this module with the following in your xmonad.hs:

import XMonad.Layout.Circle

Then edit your layoutHook by adding the Circle layout:

myLayout = Circle ||| Full ||| etc..
main = xmonad def { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see the tutorial and XMonad.Doc.Extending.

pattern Circle :: CircleEx a