XMonad.Layout.Column

Description

Provides Column layout that places all windows in one column. Windows heights are calculated from equation: H1H2 = H2H3 = ... = q, where q is given. With Shrink/Expand messages you can change the q value.

Usage

This module defines layot named Column. It places all windows in one column. Windows heights are calculated from equation: H1H2 = H2H3 = ... = q, where q is given (thus, windows heights are members of geometric progression). With Shrink/Expand messages one can change the q value.

You can use this module by adding folowing in your xmonad.hs:

import XMonad.Layout.Column

Then add layouts to your layoutHook:

myLayoutHook = Column 1.6 ||| ...

In this example, each next window will have height 1.6 times less then previous window.

newtype Column a

Constructors

Column Float