XMonad.Layout.Decoration

Description

A layout modifier and a class for easily creating decorated layouts.

Usage:

This module is intended for layout developers, who want to decorate their layouts. End users will not find here very much for them.

For examples of DecorationStyle instances you can have a look at XMonad.Layout.SimpleDecoration, XMonad.Layout.Tabbed, XMonad.Layout.DwmStyle, or XMonad.Layout.TabBarDecoration.

decoration :: (DecorationStyle ds a, Shrinker s) => s -> Theme -> ds a -> l a -> ModifiedLayout (Decoration ds s) l a

A layout modifier that, with a Shrinker, a Theme, a DecorationStyle, and a layout, will decorate this layout according to the decoration style provided.

For some usage examples see XMonad.Layout.DecorationMadness.

data Theme

A Theme is a record of colors, font etc., to customize a DecorationStyle.

For a collection of Themes see XMonad.Util.Themes

Constructors

Theme

Fields

activeColor :: String

Color of the active window

inactiveColor :: String

Color of the inactive window

urgentColor :: String

Color of the urgent window

activeBorderColor :: String

Color of the border of the active window

inactiveBorderColor :: String

Color of the border of the inactive window

urgentBorderColor :: String

Color of the border of the urgent window

activeBorderWidth :: Dimension

Width of the border of the active window

inactiveBorderWidth :: Dimension

Width of the border of the inactive window

urgentBorderWidth :: Dimension

Width of the border of the urgent window

activeTextColor :: String

Color of the text of the active window

inactiveTextColor :: String

Color of the text of the inactive window

urgentTextColor :: String

Color of the text of the urgent window

fontName :: String

Font name

decoWidth :: Dimension

Maximum width of the decorations (if supported by the DecorationStyle)

decoHeight :: Dimension

Height of the decorations

windowTitleAddons :: [(String, Align)]

Extra text to appear in a window's title bar. Refer to for a use XMonad.Layout.ImageButtonDecoration

windowTitleIcons :: [([[Bool]], Placement)]

Extra icons to appear in a window's title bar. Inner [Bool] is a row in a icon bitmap.

def :: Default a => a

The default value for this type.

data Decoration ds s a

The Decoration LayoutModifier. This data type is an instance of the LayoutModifier class. This data type will be passed, together with a layout, to the ModifiedLayout type constructor to modify the layout by adding decorations according to a DecorationStyle.

newtype DecorationMsg

A Decoration layout modifier will handle SetTheme, a message to dynamically change the decoration Theme.

Constructors

SetTheme Theme

class (Read (ds a), Show (ds a), Eq a) => DecorationStyle ds a where

The DecorationStyle class, defines methods used in the implementation of the Decoration LayoutModifier instance. A type instance of this class is passed to the Decoration type in order to decorate a layout, by using these methods.

Minimal complete definition

Nothing

Methods

describeDeco :: ds a -> String

The description that the Decoration modifier will display.

shrink :: ds a -> Rectangle -> Rectangle -> Rectangle

Shrink the window's rectangle when applying a decoration.

decorationEventHook :: ds a -> DecorationState -> Event -> X ()

The decoration event hook

decorationCatchClicksHook

Arguments

:: ds a

-> Window

-> Int
distance from the left where the click happened on the decoration

-> Int
distance from the right where the click happened on the decoration

-> X Bool

A hook that can be used to catch the cases when the user clicks on the decoration. If you return True here, the click event will be considered as dealt with and no further processing will take place.

decorationWhileDraggingHook :: ds a -> CInt -> CInt -> (Window, Rectangle) -> Position -> Position -> X ()

This hook is called while a window is dragged using the decoration. The hook can be overwritten if a different way of handling the dragging is required.

decorationAfterDraggingHook :: ds a -> (Window, Rectangle) -> Window -> X ()

This hoook is called after a window has been dragged using the decoration.

pureDecoration :: ds a -> Dimension -> Dimension -> Rectangle -> Stack a -> [(a, Rectangle)] -> (a, Rectangle) -> Maybe Rectangle

The pure version of the main method, decorate.

decorate :: ds a -> Dimension -> Dimension -> Rectangle -> Stack a -> [(a, Rectangle)] -> (a, Rectangle) -> X (Maybe Rectangle)

Given the theme's decoration width and height, the screen rectangle, the windows stack, the list of windows and rectangles returned by the underlying layout and window to be decorated, tupled with its rectangle, produce a Just Rectangle or Nothing if the window is not to be decorated.

data DefaultDecoration a

The default DecorationStyle, with just the default methods' implementations.

Constructors

DefaultDecoration

class (Read s, Show s) => Shrinker s where

Methods

shrinkIt :: s -> String -> [String]

data DefaultShrinker

shrinkText :: DefaultShrinker

data CustomShrink

Constructors

CustomShrink

shrinkWhile :: (String -> [String]) -> (String -> X Bool) -> String -> X String

isInStack :: Eq a => Stack a -> a -> Bool

True if the window is in the Stack. The Window comes second to facilitate list processing, even though w `isInStack` s won't work...;)

isVisible :: Rectangle -> [Rectangle] -> Bool

Given a Rectangle and a list of Rectangles is True if the Rectangle is not completely contained by any Rectangle of the list.

isInvisible :: Rectangle -> [Rectangle] -> Bool

The contrary of isVisible.

isWithin :: Rectangle -> Rectangle -> Bool

True is the first Rectangle is totally within the second Rectangle.

fi :: (Integral a, Num b) => a -> b

Short for fromIntegral.

findWindowByDecoration :: Window -> DecorationState -> Maybe (OrigWin, (Window, Maybe Rectangle))

module XMonad.Layout.LayoutModifier

data DecorationState

The Decoration state component, where the list of decorated window's is zipped with a list of decoration. A list of decoration is a list of tuples, a Maybe Window and a 'Maybe Rectangle'. The Window will be displayed only if the rectangle is of type Just.

type OrigWin = (Window, Rectangle)