XMonad.Layout.DecorationAddons

Description

Various stuff that can be added to the decoration. Most of it is intended to be used by other modules. See XMonad.Layout.ButtonDecoration for a module that makes use of this.

Documentation

titleBarButtonHandler :: Window -> Int -> Int -> X Bool

A function intended to be plugged into the decorationCatchClicksHook of a decoration. It will intercept clicks on the buttons of the decoration and invoke the associated action. To actually see the buttons, you will need to use a theme that includes them. See defaultThemeWithButtons below.

defaultThemeWithButtons :: Theme

Intended to be used together with titleBarButtonHandler. See above.

handleScreenCrossing :: Window -> Window -> X Bool

A function intended to be plugged into the decorationAfterDraggingHook of a decoration. It will check if the window has been dragged onto another screen and shift it there. The PositionStore is also updated accordingly, as this is designed to be used together with XMonad.Layout.PositionStoreFloat.