XMonad.Layout.DraggingVisualizer

Description

A helper module to visualize the process of dragging a window by making it follow the mouse cursor. See XMonad.Layout.WindowSwitcherDecoration for a module that makes use of this.

Documentation

draggingVisualizer :: LayoutClass l Window => l Window -> ModifiedLayout DraggingVisualizer l Window

data DraggingVisualizerMsg

Constructors

DraggingWindow Window Rectangle

DraggingStopped

data DraggingVisualizer a