XMonad.Layout.Hidden

Description

Similar to XMonad.Layout.Minimize but completely removes windows from the window set so XMonad.Layout.BoringWindows isn't necessary. Perfect companion to XMonad.Layout.BinarySpacePartition since it can be used to move windows to another part of the BSP tree.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import XMonad.Layout.Hidden

Then edit your layoutHook by adding the HiddenWindows layout modifier:

myLayout = hiddenWindows (Tall 1 (3/100) (1/2)) ||| Full ||| etc..
main = xmonad def { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

In the key bindings, do something like:

       , ((modMask, xK_backslash), withFocused hideWindow)
       , ((modMask .|. shiftMask, xK_backslash), popOldestHiddenWindow)
       ...

For detailed instruction on editing the key bindings see:

XMonad.Doc.Extending.

data HiddenWindows a

data HiddenMsg

Messages for the HiddenWindows layout modifier.

Constructors

HideWindow Window
Hide a window.

PopNewestHiddenWindow
Restore window (FILO).

PopOldestHiddenWindow
Restore window (FIFO).

PopSpecificHiddenWindow Window
Restore specific window.

hiddenWindows :: LayoutClass l Window => l Window -> ModifiedLayout HiddenWindows l Window

Apply the HiddenWindows layout modifier.

hideWindow :: Window -> X ()

Remove the given window from the current layout. It is placed in list of hidden windows so it can be restored later.

popOldestHiddenWindow :: X ()

Restore a previously hidden window. Using this function will treat the list of hidden windows as a FIFO queue. That is, the first window hidden will be restored.

popNewestHiddenWindow :: X ()

Restore a previously hidden window. Using this function will treat the list of hidden windows as a FILO queue. That is, the most recently hidden window will be restored.

popHiddenWindow :: Window -> X ()