XMonad.Layout.HintedGrid

Description

A not so simple layout that attempts to put all windows in a square grid while obeying their size hints.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Layout.HintedGrid

Then edit your layoutHook by adding the Grid layout:

 myLayout = Grid False ||| Full ||| etc..
 main = xmonad defaultConfig { layoutHook = myLayout }

You can also specify an aspect ratio for Grid to strive for with the GridRatio constructor:

 myLayout = GridRatio (4/3) False ||| etc.

For more detailed instructions on editing the layoutHook see XMonad.Doc.Extending.

data Grid a

Automatic mirroring of hinted layouts doesn't work very well, so this Grid comes with built-in mirroring. Grid False is the normal layout, Grid True is the mirrored variant (rotated by 90 degrees).

Constructors

Grid Bool

GridRatio Double Bool

arrange :: Double -> Bool -> Rectangle -> [Window] -> X [(Window, Rectangle)]

The internal function for computing the grid layout.

defaultRatio :: Double