XMonad.Layout.HintedTile

Description

A gapless tiled layout that attempts to obey window size hints, rather than simply ignoring them.

Usage

You can use this module with the following in your xmonad.hs:

import XMonad.Layout.HintedTile

Then edit your layoutHook by adding the HintedTile layout:

myLayout = hintedTile Tall ||| hintedTile Wide ||| Full ||| etc..
  where
     hintedTile = HintedTile nmaster delta ratio TopLeft
     nmaster    = 1
     ratio      = 1/2
     delta      = 3/100
main = xmonad def { layoutHook = myLayout }

Because both Xmonad and Xmonad.Layout.HintedTile define Tall, you need to disambiguate Tall. If you are replacing the built-in Tall with HintedTile, change import Xmonad to import Xmonad hiding (Tall).

For more detailed instructions on editing the layoutHook see the tutorial and XMonad.Doc.Extending.

data HintedTile a

Constructors

HintedTile

Fields

nmaster :: !Int

number of windows in the master pane

delta :: !Rational

how much to change when resizing

frac :: !Rational

ratio between master/nonmaster panes

alignment :: !Alignment

Where to place windows that are smaller than their preordained rectangles.

orientation :: !Orientation

Tall or Wide (mirrored) layout?

data Orientation

Constructors

Wide
Lay out windows similarly to Mirror tiled.

Tall
Lay out windows similarly to tiled.

data Alignment

Constructors

TopLeft

Center

BottomRight