XMonad.Layout.ImageButtonDecoration

Description

A decoration that includes small image buttons on both ends which invoke various actions when clicked on: Show a window menu (see XMonad.Actions.WindowMenu), minimize, maximize or close the window.

Note: For maximizing and minimizing to actually work, you will need to integrate XMonad.Layout.Maximize and XMonad.Layout.Minimize into your setup. See the documentation of those modules for more information.

Usage:

You can use this module with the following in your xmonad.hs:

import XMonad.Layout.ImageButtonDecoration

Then edit your layoutHook by adding the ImageButtonDecoration to your layout:

myL = imageButtonDeco shrinkText defaultThemeWithImageButtons (layoutHook def)
main = xmonad def { layoutHook = myL }

imageButtonDeco :: (Eq a, Shrinker s) => s -> Theme -> l a -> ModifiedLayout (Decoration ImageButtonDecoration s) l a

defaultThemeWithImageButtons :: Theme

shrinkText :: DefaultShrinker

data CustomShrink

Constructors

CustomShrink

class (Read s, Show s) => Shrinker s

Minimal complete definition

shrinkIt

imageTitleBarButtonHandler :: Window -> Int -> Int -> X Bool

A function intended to be plugged into the decorationCatchClicksHook of a decoration. It will intercept clicks on the buttons of the decoration and invoke the associated action. To actually see the buttons, you will need to use a theme that includes them. See defaultThemeWithImageButtons below.

data ImageButtonDecoration a