XMonad.Layout.LayoutBuilderP

Description

A layout combinator that sends windows matching given predicate to one rectangle and the rest to another.

Documentation

data LayoutP p l1 l2 a

Data type for our layout.

Constructors

LayoutP (Maybe a) (Maybe a) p SubBox (Maybe SubBox) (l1 a) (Maybe (l2 a))

layoutP

Arguments

:: (Read a, Eq a, LayoutClass l1 a, LayoutClass l2 a, LayoutClass l3 a, Predicate p a)

=> p

-> SubBox
The box to place the windows in

-> Maybe SubBox
Possibly an alternative box that is used when this layout handles all windows that are left

-> l1 a
The layout to use in the specified area

-> LayoutP p l2 l3 a
Where to send the remaining windows

-> LayoutP p l1 (LayoutP p l2 l3) a
The resulting layout

Use the specified layout in the described area windows that match given predicate and send the rest of the windows to the next layout in the chain. It is possible to supply an alternative area that will then be used instead, if there are no windows to send to the next layout.

layoutAll

Arguments

:: forall l1 p a . (Read a, Eq a, LayoutClass l1 a, Predicate p a)

=> SubBox
The box to place the windows in

-> l1 a
The layout to use in the specified area

-> LayoutP p l1 Full a
The resulting layout

Use the specified layout in the described area for all remaining windows.

relBox

Arguments

:: Rational
Relative X-Position with respect to the surrounding area

-> Rational
Relative Y-Position with respect to the surrounding area

-> Rational
Relative width with respect to the remaining width

-> Rational
Relative height with respect to the remaining height

-> SubBox
The resulting SubBox describing the area

Create a box with only relative measurements.

absBox

Arguments

:: Int
Absolute X-Position

-> Int
Absolute Y-Position

-> Int
Absolute width

-> Int
Absolute height

-> SubBox
The resulting SubBox describing the area

Create a box with only absolute measurements. If the values are negative, the total remaining space will be added. For sizes it will also be added for zeroes.

Overloading ways to select windows

Predicate exists because layouts are required to be serializable, and XMonad.Util.WindowProperties is not sufficient (for example it does not allow using regular expressions).

compare XMonad.Util.Invisible

class Predicate p w where

Type class for predicates. This enables us to manage not only Windows, but any objects, for which instance Predicate is defined.

Another instance exists in XMonad.Util.WindowPropertiesRE in xmonad-extras

Methods

alwaysTrue

Arguments

:: Proxy w

-> p
A predicate that is always True.

checkPredicate

Arguments

:: p

-> w

-> X Bool
Check if given object (window or smth else) matches that predicate

data Proxy a

Contains no actual data, but is needed to help select the correct instance of Predicate

Constructors

Proxy