XMonad.Layout.LimitWindows

Description

A layout modifier that limits the number of windows that can be shown. See XMonad.Layout.Minimize for manually setting hidden windows.

Usage

To use this module, add the following import to xmonad.hs:

import XMonad.Layout.LimitWindows

myLayout = limitWindows 6 $ Tall 1 0.03 0.5 ||| Full ||| RandomOtherLayout...
main = xmonad def { layoutHook = myLayout }

You may also be interested in dynamically changing the number dynamically, by binding keys to the increaseLimit, decreaseLimit, or setLimit actions.

For detailed instructions on editing your key bindings, see the tutorial.

See also boringAuto for keybindings that skip the hidden windows.

Layout Modifiers

limitWindows :: Int -> l a -> ModifiedLayout LimitWindows l a

Only display the first n windows.

limitSlice :: Int -> l a -> ModifiedLayout LimitWindows l a

Only display n windows around the focused window. This makes sense with layouts that arrange windows linearily, like XMonad.Layout.Accordion.

limitSelect :: Int -> Int -> l a -> ModifiedLayout Selection l a

Only display the first m windows and r others. The IncMasterN message will change m, as well as passing it onto the underlying layout.

Change the number of windows

increaseLimit :: X ()

decreaseLimit :: X ()

setLimit :: Int -> X ()

Types

data LimitWindows a

data Selection a