XMonad.Layout.Master

Description

Layout modfier that adds a master window to another layout.

Usage

You can use this module with the following in your xmonad.hs:

import XMonad.Layout.Master

Then edit your layoutHook and add the Master modifier to the layout that you prefer.

mastered (1/100) (1/2) $ Grid

Or if you prefer to have a master with fixed width:

fixMastered (1/100) (1/2) $ Grid

Or if you want multiple (here two) master windows from the beginning:

multimastered 2 (1/100) (1/2) $ Grid

This will use the left half of your screen for a master window and let Grid manage the right half.

For more detailed instructions on editing the layoutHook see the tutorial and XMonad.Doc.Extending.

Like Tall, withMaster supports the Shrink and XMonad.Layout.Expand' messages.

mastered

Arguments

:: LayoutClass l a

=> Rational
delta, the ratio of the screen to resize by

-> Rational
frac, what portion of the screen to use for the master window

-> l a
the layout to be modified

-> ModifiedLayout AddMaster l a

fixMastered

Arguments

:: LayoutClass l a

=> Rational
delta, the ratio of the screen to resize by

-> Rational
frac, what portion of the screen to use for the master window

-> l a
the layout to be modified

-> ModifiedLayout FixMaster l a

multimastered

Arguments

:: LayoutClass l a

=> Int
k, number of master windows

-> Rational
delta, the ratio of the screen to resize by

-> Rational
frac, what portion of the screen to use for the master window

-> l a
the layout to be modified

-> ModifiedLayout AddMaster l a

data AddMaster a

Data type for LayoutModifier which converts given layout to a mastered layout