XMonad.Layout.Maximize

Description

Temporarily yanks the focused window out of the layout to mostly fill the screen.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Layout.Maximize

Then edit your layoutHook by adding the Maximize layout modifier:

 myLayout = maximize (Tall 1 (3/100) (1/2)) ||| Full ||| etc..
 main = xmonad defaultConfig { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

In the key-bindings, do something like:

        , ((modm, xK_backslash), withFocused (sendMessage . maximizeRestore))
        ...

For detailed instruction on editing the key binding see:

XMonad.Doc.Extending.

maximize :: LayoutClass l Window => l Window -> ModifiedLayout Maximize l Window

maximizeRestore :: Window -> MaximizeRestore

data Maximize a

data MaximizeRestore