XMonad.Layout.Minimize

Description

Makes it possible to minimize windows, temporarily removing them from the layout until they are restored.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Layout.Minimize

Then edit your layoutHook by adding the Minimize layout modifier:

 myLayout = minimize (Tall 1 (3/100) (1/2)) ||| Full ||| etc..
 main = xmonad defaultConfig { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

In the key-bindings, do something like:

        , ((modm,               xK_m     ), withFocused minimizeWindow)
        , ((modm .|. shiftMask, xK_m     ), sendMessage RestoreNextMinimizedWin)

The first action will minimize the focused window, while the second one will restore the next minimized window.

For detailed instruction on editing the key binding see:

XMonad.Doc.Extending.

The module is designed to work together with XMonad.Layout.BoringWindows so that minimized windows will be skipped over when switching the focused window with the keyboard. Include boringWindows in your layout hook and see the documentation of XMonad.Layout.BoringWindows on how to modify your keybindings.

Also see XMonad.Hooks.Minimize if you want to be able to minimize and restore windows from your taskbar.

minimize :: LayoutClass l Window => l Window -> ModifiedLayout Minimize l Window

minimizeWindow :: Window -> X ()

data MinimizeMsg

Constructors

RestoreMinimizedWin Window

RestoreNextMinimizedWin

data Minimize a