XMonad.Layout.NoFrillsDecoration

Description

Most basic version of decoration for windows without any additional modifications. In contrast to XMonad.Layout.SimpleDecoration this will result in title bars that span the entire window instead of being only the length of the window title.

Usage:

You can use this module with the following in your xmonad.hs:

import XMonad.Layout.NoFrillsDecoration

Then edit your layoutHook by adding the NoFrillsDecoration to your layout:

myL = noFrillsDeco shrinkText def (layoutHook def)
main = xmonad def { layoutHook = myL }

noFrillsDeco :: (Eq a, Shrinker s) => s -> Theme -> l a -> ModifiedLayout (Decoration NoFrillsDecoration s) l a

Add very simple decorations to windows of a layout.

module XMonad.Layout.SimpleDecoration

data NoFrillsDecoration a