XMonad.Layout.OneBig

Description

Provides layout named OneBig. It places one (master) window at top left corner of screen, and other (slave) windows at top

Usage

This module defines layout named OneBig. It places one (master) window at top left, and other (slave) windows at right and at bottom of master. It tries to give equal space for each slave window.

You can use this module by adding following in your xmonad.hs:

 import XMonad.Layout.OneBig

Then add layouts to your layoutHook:

 myLayoutHook = OneBig (3/4) (3/4) ||| ...

In this example, master window will occupy 3/4 of screen width and 3/4 of screen height.

data OneBig a

Data type for layout

Constructors

OneBig Float Float