XMonad.Layout.Renamed

Description

Layout modifier that can modify the description of its underlying layout on a (hopefully) flexible way.

Usage

You can use this module by adding

import XMonad.Layout.Renamed

to your xmonad.hs.

You can then use renamed to modify the description of your layouts. For example:

myLayout = renamed [PrependWords "Awesome"] $ tiled ||| Mirror tiled ||| Full

renamed :: [Rename a] -> l a -> ModifiedLayout Rename l a

Apply a list of Rename values to a layout, from left to right.

named :: String -> l a -> ModifiedLayout Rename l a

Rename a layout. (Convenience alias for renamed [Replace s].)

data Rename a

The available renaming operations

Constructors

CutLeft Int
Remove a number of characters from the left

CutRight Int
Remove a number of characters from the right

Append String
Add a string on the right

Prepend String
Add a string on the left

CutWordsLeft Int
Remove a number of words from the left

CutWordsRight Int
Remove a number of words from the right

KeepWordsLeft Int
Keep a number of words from the left

KeepWordsRight Int
Keep a number of words from the right

AppendWords String
Add a string to the right, prepending a space to it if necessary

PrependWords String
Add a string to the left, appending a space to it if necessary

Replace String
Replace with another string

Chain [Rename a]
Apply a list of modifications in left-to-right order