XMonad.Layout.ResizableThreeColumns

Description

A layout similar to tall but with three columns. With 2560x1600 pixels this layout can be used for a huge main window and up to six reasonable sized resizable slave windows.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import XMonad.Layout.ResizableThreeColumns

Then edit your layoutHook by adding the ResizableThreeCol layout:

myLayout = ResizableThreeCol 1 (3/100) (1/2) [] ||| ResizableThreeColMid 1 (3/100) (1/2) [] ||| etc..
main = xmonad def { layoutHook = myLayout }

The first argument specifies how many windows initially appear in the main window. The second argument argument specifies the amount to resize while resizing and the third argument specifies the initial size of the columns. A positive size designates the fraction of the screen that the main window should occupy, but if the size is negative the absolute value designates the fraction a slave column should occupy. If both slave columns are visible, they always occupy the same amount of space.

You may also want to add the following key bindings:

, ((modm,               xK_a), sendMessage MirrorShrink)
, ((modm,               xK_z), sendMessage MirrorExpand)

The ResizableThreeColMid variant places the main window between the slave columns.

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

data ResizableThreeCol a

Arguments are nmaster, delta, fraction

Constructors

ResizableThreeColMid

Fields

threeColNMaster :: !Int

threeColDelta :: !Rational

threeColFrac :: !Rational

threeColSlaves :: [Rational]

ResizableThreeCol

Fields

threeColNMaster :: !Int

threeColDelta :: !Rational

threeColFrac :: !Rational

threeColSlaves :: [Rational]

data MirrorResize

Constructors

MirrorShrink

MirrorExpand