XMonad.Layout.ResizableTile

Description

More useful tiled layout that allows you to change a width/height of window.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Layout.ResizableTile

Then edit your layoutHook by adding the ResizableTile layout:

 myLayout =  ResizableTall 1 (3/100) (1/2) [] ||| etc..
 main = xmonad defaultConfig { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

You may also want to add the following key bindings:

 , ((modm,               xK_a), sendMessage MirrorShrink)
 , ((modm,               xK_z), sendMessage MirrorExpand)

For detailed instruction on editing the key binding see:

XMonad.Doc.Extending.

data ResizableTall a

Constructors

ResizableTall

Fields

_nmaster :: Int

number of master windows

_delta :: Rational

change when resizing by Shrink, Expand, MirrorShrink, MirrorExpand

_frac :: Rational

width of master

_slaves :: [Rational]

fraction to multiply the window height that would be given when divided equally.

slave windows are assigned their modified heights in order, from top to bottom

unspecified values are replaced by 1

data MirrorResize

Constructors

MirrorShrink

MirrorExpand