XMonad.Layout.ResizeScreen

Description

A layout transformer to have a layout respect a given screen geometry. Mostly used with XMonad.Layout.Decoration (the Horizontal and the Vertical version will react to SetTheme and change their dimension accordingly.

Usage:

You can use this module by importing it into your xmonad.hs file:

import XMonad.Layout.ResizeScreen

and modifying your layoutHook as follows (for example):

layoutHook = resizeHorizontal 40 Full

For more detailed instructions on editing the layoutHook see the tutorial and XMonad.Doc.Extending.

resizeHorizontal :: Int -> l a -> ModifiedLayout ResizeScreen l a

resizeVertical :: Int -> l a -> ModifiedLayout ResizeScreen l a

resizeHorizontalRight :: Int -> l a -> ModifiedLayout ResizeScreen l a

resizeVerticalBottom :: Int -> l a -> ModifiedLayout ResizeScreen l a

withNewRectangle :: Rectangle -> l a -> ModifiedLayout ResizeScreen l a

data ResizeScreen a

Constructors

ResizeScreen ResizeMode Int

WithNewScreen Rectangle

data ResizeMode