XMonad.Layout.ShowWName

Description

This is a layout modifier that will show the workspace name

Usage

You can use this module with the following in your xmonad.hs:

import XMonad.Layout.ShowWName
myLayout = layoutHook def
main = xmonad def { layoutHook = showWName myLayout }

For more detailed instructions on editing the layoutHook see the tutorial and XMonad.Doc.Extending.

showWName :: l a -> ModifiedLayout ShowWName l a

A layout modifier to show the workspace name when switching

showWName' :: SWNConfig -> l a -> ModifiedLayout ShowWName l a

A layout modifier to show the workspace name when switching. It is possible to provide a custom configuration.

def :: Default a => a

The default value for this type.

data SWNConfig

Constructors

SWNC

Fields

swn_font :: String

Font name

swn_bgcolor :: String

Background color

swn_color :: String

String color

swn_fade :: Rational

Time in seconds of the name visibility

data ShowWName a