XMonad.Layout.ShowWName

Description

This is a layout modifier that will show the workspace name

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Layout.ShowWName
 myLayout = layoutHook defaultConfig
 main = xmonad defaultConfig { layoutHook = showWName myLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

showWName :: l a -> ModifiedLayout ShowWName l a

A layout modifier to show the workspace name when switching

showWName' :: SWNConfig -> l a -> ModifiedLayout ShowWName l a

A layout modifier to show the workspace name when switching. It is possible to provide a custom configuration.

defaultSWNConfig :: SWNConfig

data SWNConfig

Constructors

SWNC

Fields

swn_font :: String

Font name

swn_bgcolor :: String

Background color

swn_color :: String

String color

swn_fade :: Rational

Time in seconds of the name visibility

data ShowWName a