XMonad.Layout.SimpleFloat

Description

A basic floating layout.

Usage:

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Layout.SimpleFloat

Then edit your layoutHook by adding the SimpleFloat layout:

 myLayout = simpleFloat ||| Full ||| etc..
 main = xmonad defaultConfig { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

simpleFloat :: Eq a => ModifiedLayout (Decoration SimpleDecoration DefaultShrinker) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger SimpleFloat)) a

A simple floating layout where every window is placed according to the window's initial attributes.

This version is decorated with the SimpleDecoration style.

simpleFloat' :: (Eq a, Shrinker s) => s -> Theme -> ModifiedLayout (Decoration SimpleDecoration s) (ModifiedLayout MouseResize (ModifiedLayout WindowArranger SimpleFloat)) a

Same as simpleFloat, but with the possibility of setting a custom shrinker and a custom theme.

data SimpleDecoration a

Constructors

Simple Bool

data SimpleFloat a

Constructors

SF Dimension

shrinkText :: DefaultShrinker

data CustomShrink

Constructors

CustomShrink

class (Read s, Show s) => Shrinker s where

Methods

shrinkIt :: s -> String -> [String]