XMonad.Layout.SortedLayout

Description

A LayoutModifier that sorts the windows in another layout, given a list of properties. The order of properties in the list determines the order of windows in the final layout. Any unmatched windows go to the end of the order.

Usage:

You can use this module with the following in your xmonad.hs:

import XMonad.Layout.SortedLayout

Then edit your layoutHook to sort another layout (in this case, Grid):

myLayout = sorted [ClassName "Firefox", ClassName "URxvt"] Grid
main = xmonad def { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see the tutorial and XMonad.Doc.Extending.

sorted :: [Property] -> l a -> ModifiedLayout SortedLayout l a

Modify a layout using a list of properties to sort its windows.

data Property

Most of the property constructors are quite self-explaining.

Constructors

Title String

ClassName String

Resource String

Role String
WM_WINDOW_ROLE property

Machine String
WM_CLIENT_MACHINE property

And Property Property infixr 9

Or Property Property infixr 8

Not Property

Const Bool

Tagged String
Tagged via XMonad.Actions.TagWindows