XMonad.Layout.Spacing

Description

Add a configurable amount of space around windows.

Usage

You can use this module by importing it into your ~/.xmonad/xmonad.hs file:

import XMonad.Layout.Spacing

and modifying your layoutHook as follows (for example):

layoutHook = spacing 2 $ Tall 1 (3/100) (1/2)
                     -- put a 2px space around every window

spacing :: Int -> l a -> ModifiedLayout Spacing l a

Surround all windows by a certain number of pixels of blank space.

data Spacing a

spacingWithEdge :: Int -> l a -> ModifiedLayout SpacingWithEdge l a

Surround all windows by a certain number of pixels of blank space, and additionally adds the same amount of spacing around the edge of the screen.

data SpacingWithEdge a

smartSpacing :: Int -> l a -> ModifiedLayout SmartSpacing l a

Surrounds all windows with blank space, except when the window is the only visible window on the current workspace.

data SmartSpacing a

smartSpacingWithEdge :: Int -> l a -> ModifiedLayout SmartSpacingWithEdge l a

Surrounds all windows with blank space, and adds the same amount of spacing around the edge of the screen, except when the window is the only visible window on the current workspace.

data SmartSpacingWithEdge a

data ModifySpacing

Message to dynamically modify (e.g. increasedecreaseset) the size of the window spacing

Constructors

ModifySpacing (Int -> Int)

setSpacing :: Int -> X ()

Set spacing to given amount

incSpacing :: Int -> X ()

Increase spacing by given amount