XMonad.Layout.Spiral

Description

A spiral tiling layout.

Usage

You can use this module with the following in your xmonad.hs:

import XMonad.Layout.Spiral

Then edit your layoutHook by adding the Spiral layout:

myLayout =  spiral (6/7) ||| etc..
main = xmonad def { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see the tutorial and XMonad.Doc.Extending.

spiral :: Rational -> SpiralWithDir a

A spiral layout. The parameter controls the size ratio between successive windows in the spiral. Sensible values range from 0 up to the aspect ratio of your monitor (often 4/3).

By default, the spiral is counterclockwise, starting to the east. See also spiralWithDir.

spiralWithDir :: Direction -> Rotation -> Rational -> SpiralWithDir a

Create a spiral layout, specifying the starting cardinal direction, the spiral direction (clockwise or counterclockwise), and the size ratio.

data Rotation

Constructors

CW

CCW

data Direction

Constructors

East

South

West

North

data SpiralWithDir a