XMonad.Layout.Square

Description

A layout that splits the screen into a square area and the rest of the screen. This is probably only ever useful in combination with XMonad.Layout.Combo. It sticks one window in a square region, and makes the rest of the windows live with what's left (in a full-screen sense).

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs file:

   import XMonad.Layout.Square

An example layout using square together with XMonad.Layout.Combo to make the very last area square:

 , combo (combo (mirror $ twoPane 0.03 0.85),1)] (twoPane 0.03 0.5) )
                [(twoPane 0.03 0.2,1),(combo [(twoPane 0.03 0.8,1),(square,1)]
         [(tabbed,3),(tabbed,30),(tabbed,1),(tabbed,1)]

data Square a

Constructors

Square