XMonad.Layout.TabBarDecoration

Description

A layout modifier to add a bar of tabs to your layouts.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Layout.TabBarDecoration

Then edit your layoutHook by adding the layout you want:

 main = xmonad defaultConfig { layoutHook = simpleTabBar $ layoutHook defaultConfig}

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

tabBar will give you the possibility of setting a custom shrinker and a custom theme.

The deafult theme can be dynamically change with the xmonad theme selector. See XMonad.Prompt.Theme. For more themse, look at XMonad.Util.Themes

simpleTabBar :: Eq a => l a -> ModifiedLayout (Decoration TabBarDecoration DefaultShrinker) (ModifiedLayout ResizeScreen l) a

Add, on the top of the screen, a simple bar of tabs to a given | layout, with the default theme and the default shrinker.

tabBar :: (Eq a, Shrinker s) => s -> Theme -> XPPosition -> l a -> ModifiedLayout (Decoration TabBarDecoration s) l a

Same of simpleTabBar, but with the possibility of setting a custom shrinker, a custom theme and the position: Top or Bottom.

defaultTheme :: Theme

The default xmonad Theme.

shrinkText :: DefaultShrinker

data TabBarDecoration a

Constructors

TabBar XPPosition

data XPPosition

Constructors

Top

Bottom

module XMonad.Layout.ResizeScreen