XMonad.Layout.ToggleLayouts

Description

A module to toggle between two layouts.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

import XMonad.Layout.ToggleLayouts

Then edit your layoutHook by adding the ToggleLayouts layout:

myLayout = toggleLayouts Full (Tall 1 (3/100) (1/2)) ||| etc..
main = xmonad def { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

To toggle between layouts add a key binding like

   , ((modm .|. controlMask, xK_space), sendMessage ToggleLayout)

or a key binding like

   , ((modm .|. controlMask, xK_space), sendMessage (Toggle "Full"))

For detailed instruction on editing the key binding see:

XMonad.Doc.Extending.

toggleLayouts :: (LayoutClass lt a, LayoutClass lf a) => lt a -> lf a -> ToggleLayouts lt lf a

data ToggleLayout

Constructors

ToggleLayout

Toggle String

data ToggleLayouts lt lf a