XMonad.Layout.WindowNavigation

Description

WindowNavigation is an extension to allow easy navigation of a workspace.

Usage

You can use this module with the following in your xmonad.hs:

import XMonad.Layout.WindowNavigation

Then edit your layoutHook by adding the WindowNavigation layout modifier to some layout:

myLayout = windowNavigation (Tall 1 (3/100) (1/2)) ||| Full ||| etc..
main = xmonad def { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see the tutorial and XMonad.Doc.Extending.

In keybindings:

   , ((modm,                 xK_Right), sendMessage $ Go R)
   , ((modm,                 xK_Left ), sendMessage $ Go L)
   , ((modm,                 xK_Up   ), sendMessage $ Go U)
   , ((modm,                 xK_Down ), sendMessage $ Go D)
   , ((modm .|. controlMask, xK_Right), sendMessage $ Swap R)
   , ((modm .|. controlMask, xK_Left ), sendMessage $ Swap L)
   , ((modm .|. controlMask, xK_Up   ), sendMessage $ Swap U)
   , ((modm .|. controlMask, xK_Down ), sendMessage $ Swap D)

For detailed instruction on editing the key binding see:

the tutorial.

windowNavigation :: LayoutClass l a => l a -> ModifiedLayout WindowNavigation l a

configurableNavigation :: LayoutClass l a => WNConfig -> l a -> ModifiedLayout WindowNavigation l a

data Navigate

Constructors

Go Direction2D

Swap Direction2D

Move Direction2D

Apply (Window -> X ()) Direction2D
Apply action with destination window

data Direction2D

Two-dimensional directions:

Constructors

U
Up

D
Down

R
Right

L
Left

data MoveWindowToWindow a

Constructors

MoveWindowToWindow a a

data WNConfig

Used with configurableNavigation to specify how to show reachable windows' borders. You cannot create WNConfig values directly; use def or one of the following three functions to create one.

def, and windowNavigation, uses the focused border color at 40% brightness, as if you had specified

configurableNavigation (navigateBrightness 0.4)

navigateColor :: String -> WNConfig

Indicate reachable windows by drawing their borders in the specified color.

navigateBrightness :: Double -> WNConfig

Indicate reachable windows by drawing their borders in the active border color, with the specified brightness.

noNavigateBorders :: WNConfig

Don't use window borders for navigation.

def :: Default a => a

The default value for this type.

data WindowNavigation a