XMonad.Layout.WorkspaceDir

Description

WorkspaceDir is an extension to set the current directory in a workspace.

Actually, it sets the current directory in a layout, since there's no way I know of to attach a behavior to a workspace. This means that any terminals (or other programs) pulled up in that workspace (with that layout) will execute in that working directory. Sort of handy, I think.

Note this extension requires the directory package to be installed.

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Layout.WorkspaceDir

Then edit your layoutHook by adding the Workspace layout modifier to some layout:

 myLayout = workspaceDir "~" (Tall 1 (3/100) (1/2))  ||| Full ||| etc..
 main = xmonad defaultConfig { layoutHook = myLayout }

For more detailed instructions on editing the layoutHook see:

XMonad.Doc.Extending

WorkspaceDir provides also a prompt. To use it you need to import XMonad.Prompt and add something like this to your key bindings:

  , ((modm .|. shiftMask, xK_x     ), changeDir defaultXPConfig)

For detailed instruction on editing the key binding see:

XMonad.Doc.Extending.

workspaceDir :: LayoutClass l a => String -> l a -> ModifiedLayout WorkspaceDir l a

changeDir :: XPConfig -> X ()

data WorkspaceDir a