XMonad.Prelude

Description

Utility functions and re-exports for a more ergonomic developing experience. Users themselves will not find much use here.

Documentation

module Data.Traversable

module Data.Monoid

module Data.Maybe

module Data.List

module Data.Functor

module Data.Function

module Data.Foldable

module Data.Char

module Data.Bool

module Control.Monad

module Control.Applicative

fi :: (Integral a, Num b) => a -> b

Short for fromIntegral.

chunksOf :: Int -> [a] -> [[a]]

Given a maximum length, splits a list into sublists

>>> chunksOf 5 (take 30 $ repeat 'a')
["aaaaa","aaaaa","aaaaa","aaaaa","aaaaa","aaaaa"]

(.:) :: (a -> b) -> (c -> d -> a) -> c -> d -> b

Multivariable composition.

f .: g ≡ (f .) . g ≡ \c d -> f (g c d)

(!?) :: [a] -> Int -> Maybe a

Safe version of (!!).

data NonEmpty a

Non-empty (and non-strict) list type.

Since: base-4.9.0.0

Constructors

a :| [a] infixr 5

notEmpty :: HasCallStack => [a] -> NonEmpty a

fromList with a better error message. Useful to silence GHC's Pattern match(es) are non-exhaustive warning in places where the programmer knows it's always non-empty, but it's infeasible to express that in the type system.

safeGetWindowAttributes :: Window -> X (Maybe WindowAttributes)

A safe version of getWindowAttributes.

mkAbsolutePath :: MonadIO m => FilePath -> m FilePath

(Naïvely) turn a relative path into an absolute one.

If the path starts with /, do nothing.

If it starts with ~/, replace that with the actual home

directory.

If it starts with $, read the name of an environment

variable and replace it with the contents of that.

Otherwise, prepend the home directory and / to the path.

Keys

keyToString :: (KeyMask, KeySym) -> String

Convert a full key combination; i.e., a KeyMask and KeySym pair, into a string.

keymaskToString

Arguments

:: KeyMask
Num lock mask

-> KeyMask
Modifier mask

-> String

Convert a modifier mask into a useful string.

cleanKeyMask :: X (KeyMask -> KeyMask)

Strip numlock, capslock, mouse buttons and XKB group from a KeyMask, leaving only modifier keys like Shift, Control, Super, Hyper in the mask (hence the "Key" in "cleanKeyMask").

Core's cleanMask only strips the first two because key events from passive grabs (key bindings) are stripped of mouse buttons and XKB group by the X server already for compatibility reasons. For more info, see: https://www.x.org/releases/X11R7.7/doc/kbproto/xkbproto.html#Delivering_a_Key_or_Button_Event_to_a_Client

regularKeys :: [(String, KeySym)]

A list of "regular" (extended ASCII) keys.

allSpecialKeys :: [(String, KeySym)]

A list of all special key names and their associated KeySyms.

specialKeys :: [(String, KeySym)]

A list of special key names and their corresponding KeySyms.

multimediaKeys :: [(String, KeySym)]

List of multimedia keys. If Xlib does not know about some keysym it's omitted from the list (stringToKeysym returns noSymbol in this case).

functionKeys :: [(String, KeySym)]

A list pairing function key descriptor strings (e.g. "F2") with the associated KeySyms.

type WindowScreen = Screen WorkspaceId (Layout Window) Window ScreenId ScreenDetail

The specialized Screen derived from WindowSet.