XMonad.Prompt.Layout

Description

A layout-selection prompt for XMonad

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Prompt
 import XMonad.Prompt.Layout

   , ((modm .|. shiftMask, xK_m     ), layoutPrompt defaultXPConfig)

For detailed instruction on editing the key binding see XMonad.Doc.Extending.

WARNING: This prompt won't display all possible layouts, because the code to enable this was rejected from xmonad core. It only displays layouts that are actually in use. Also, you can only select layouts if you are using NewSelect, rather than the Select defined in xmonad core (which doesn't have this feature). So all in all, this module is really more a proof-of-principle than something you can actually use productively.

layoutPrompt :: XPConfig -> X ()