XMonad.Prompt.Man

Description

A manual page prompt for XMonad window manager.

TODO

narrow completions by section number, if the one is specified (like /etc/bash_completion does)

Usage

In your xmonad.hs:

import XMonad.Prompt
import XMonad.Prompt.Man

In your keybindings add something like:

    , ((modm, xK_F1), manPrompt def)

For detailed instruction on editing the key binding see the tutorial.

manPrompt :: XPConfig -> X ()

Query for manual page to be displayed.

getCommandOutput :: String -> IO String

Run a command using shell and return its output.

XXX Merge into XMonad.Util.Run?

(Ask "gurus" whether evaluate (length ...) approach is better/more idiomatic.)

data Man