XMonad.Prompt.Unicode

Description

A prompt for searching unicode characters by name and inserting them into the clipboard.

The provided unicodePrompt and typeUnicodePrompt use xsel and xdotool respectively.

Usage

You can use this module by importing it, along with XMonad.Prompt, into your xmonad.hs file:

import XMonad.Prompt
import XMonad.Prompt.Unicode

and adding an appropriate keybinding, for example:

 , ((modm .|. controlMask, xK_u), unicodePrompt "/path/to/unicode-data" def)

A path to a UnicodeData.txt file or equivalent must be provided. This file should be available through your package manager; search for unicode-data. If no package is found, one may opt to download this file directly from [unicode.org](http:/www.unicode.orgPublicUNIDATAUnicodeData.txt).

More flexibility is given by the mkUnicodePrompt function, which takes a command and a list of arguments to pass as its first two arguments. See unicodePrompt for details.

unicodePrompt :: String -> XPConfig -> X ()

Prompt the user for a Unicode character to be inserted into the paste buffer of the X server.

typeUnicodePrompt :: String -> XPConfig -> X ()

Prompt the user for a Unicode character to be typed by xdotool.

mkUnicodePrompt :: String -> [String] -> String -> XPConfig -> X ()