XMonad.Prompt.Window

Description

xprompt operations to bring windows to you, and bring you to windows.

Usage

WindowPrompt brings windows to you and you to windows. That is to say, it pops up a prompt with window names, in case you forgot where you left your XChat.

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Prompt
 import XMonad.Prompt.Window

and in the keys definition:

 , ((modm .|. shiftMask, xK_g     ), windowPromptGoto  defaultXPConfig)
 , ((modm .|. shiftMask, xK_b     ), windowPromptBring defaultXPConfig)

The autoComplete option is a handy complement here:

 , ((modm .|. shiftMask, xK_g     ), windowPromptGoto
                                            defaultXPConfig { autoComplete = Just 500000 } )

The '500000' is the number of microseconds to pause before sending you to your new window. This is useful so that you don't accidentally send some keystrokes to the selected client.

For detailed instruction on editing the key binding see XMonad.Doc.Extending.

windowPromptGoto :: XPConfig -> X ()

windowPromptBring :: XPConfig -> X ()

windowPromptBringCopy :: XPConfig -> X ()

data WindowPrompt