XMonad.Prompt.Workspace

Description

A workspace prompt for XMonad

Usage

You can use this module with the following in your ~/.xmonad/xmonad.hs:

 import XMonad.Prompt
 import XMonad.Prompt.Workspace

   , ((modm .|. shiftMask, xK_m     ), workspacePrompt defaultXPConfig (windows . W.shift))

For detailed instruction on editing the key binding see XMonad.Doc.Extending.

workspacePrompt :: XPConfig -> (String -> X ()) -> X ()

For developers

data Wor

Constructors

Wor String