XMonad.Util.DebugWindow

Description

Module to dump window information for diagnostic/debugging purposes. See XMonad.Hooks.DebugEvents and XMonad.Hooks.DebugStack for practical uses.

Documentation

debugWindow :: Window -> X String

Output a window by ID in hex, decimal, its ICCCM resource name and class, its title if available, and EWMH type and state if available. Also indicate override_redirect with an exclamation mark, and wrap in brackets if it is unmapped or withdrawn.