XMonad.Util.Dmenu

Description

A convenient binding to dmenu.

Requires the process-1.0 package

Usage

You can use this module with the following in your Config.hs file:

import XMonad.Util.Dmenu

These functions block xmonad's event loop until dmenu exits; this means that programs will not be able to open new windows and you will not be able to change workspaces or input focus until you have responded to the prompt one way or another.

dmenu :: MonadIO m => [String] -> m String

Run dmenu to select an option from a list.

dmenuXinerama :: [String] -> X String

Starts dmenu on the current screen. Requires this patch to dmenu: http://www.jcreigh.com/dmenu/dmenu-3.2-xinerama.patch

dmenuMap :: MonadIO m => Map String a -> m (Maybe a)

Run dmenu to select an entry from a map based on the key.

menu :: MonadIO m => String -> [String] -> m String

like dmenu but also takes the command to run.

menuArgs :: MonadIO m => String -> [String] -> [String] -> m String

Like menu but also takes a list of command line arguments.

menuMap :: MonadIO m => String -> Map String a -> m (Maybe a)

Like dmenuMap but also takes the command to run.

menuMapArgs :: MonadIO m => String -> [String] -> Map String a -> m (Maybe a)

Like menuMap but also takes a list of command line arguments.