XMonad.Util.Invisible

Description

A data type to store the layout state

Usage:

A wrapper data type to store layout state that shouldn't be persisted across restarts. A common wrapped type to use is Maybe a. Invisible derives trivial definitions for Read and Show, so the wrapped data type need not do so.

newtype Invisible m a

Constructors

I (m a)

whenIJust :: Monad m => Invisible Maybe a -> (a -> m ()) -> m ()

fromIMaybe :: a -> Invisible Maybe a -> a