XMonad.Util.PositionStore

Description

A utility module to store information about position and size of a window. See XMonad.Layout.PositionStoreFloat for a layout that makes use of this.

Documentation

getPosStore :: X PositionStore

modifyPosStore :: (PositionStore -> PositionStore) -> X ()

posStoreInsert :: PositionStore -> Window -> Rectangle -> Rectangle -> PositionStore

posStoreMove :: PositionStore -> Window -> Position -> Position -> Rectangle -> Rectangle -> PositionStore

posStoreQuery :: PositionStore -> Window -> Rectangle -> Maybe Rectangle

posStoreRemove :: PositionStore -> Window -> PositionStore

data PositionStore