XMonad.Util.StringProp

Description

Internal utility functions for storing Strings with the root window.

Used for global state like IORefs with string keys, but more latency, persistent between xmonad restarts.

Documentation

type StringProp = String

getStringProp :: MonadIO m => Display -> StringProp -> m (Maybe [Char])

Get the name of a string property and returns it as a Maybe.

setStringProp :: MonadIO m => Display -> StringProp -> [Char] -> m ()

Set the value of a string property.

getStringListProp :: MonadIO m => Display -> StringProp -> m [String]

Given a property name, returns its contents as a list. It uses the empty list as default value.

setStringListProp :: MonadIO m => Display -> StringProp -> [String] -> m ()

Given a property name and a list, sets the value of this property with the list given as argument.