XMonad.Util.Timer

Description

A module for setting up timers

Usage

This module can be used to setup a timer to handle deferred events. See ShowWName for an usage example.

startTimer :: Rational -> X TimerId

Start a timer, which will send a ClientMessageEvent after some time (in seconds).

handleTimer :: TimerId -> Event -> X (Maybe a) -> X (Maybe a)

Given a TimerId and an Event, run an action when the Event has been sent by the timer specified by the TimerId

type TimerId = Int