XMonad.Util.XUtils

Description

A module for painting on the screen

Usage:

See XMonad.Layout.Tabbed or XMonad.Layout.DragPane or XMonad.Layout.Decoration for usage examples

averagePixels :: Pixel -> Pixel -> Double -> X Pixel

Compute the weighted average the colors of two given Pixel values.

createNewWindow :: Rectangle -> Maybe EventMask -> String -> Bool -> X Window

Create a simple window given a rectangle. If Nothing is given only the exposureMask will be set, otherwise the Just value. Use showWindow to map and hideWindow to unmap.

showWindow :: Window -> X ()

Map a window

showWindows :: [Window] -> X ()

the list version

hideWindow :: Window -> X ()

unmap a window

hideWindows :: [Window] -> X ()

the list version

deleteWindow :: Window -> X ()

destroy a window

deleteWindows :: [Window] -> X ()

the list version

paintWindow

Arguments

:: Window
The window where to draw

-> Dimension
Window width

-> Dimension
Window height

-> Dimension
Border width

-> String
Window background color

-> String
Border color

-> X ()

Fill a window with a rectangle and a border

paintAndWrite

Arguments

:: Window
The window where to draw

-> XMonadFont
XMonad Font for drawing

-> Dimension
Window width

-> Dimension
Window height

-> Dimension
Border width

-> String
Window background color

-> String
Border color

-> String
String color

-> String
String background color

-> [Align]
String Alignments

-> [String]
Strings to be printed

-> X ()

Fill a window with a rectangle and a border, and write | a number of strings to given positions

paintTextAndIcons

Arguments

:: Window
The window where to draw

-> XMonadFont
XMonad Font for drawing

-> Dimension
Window width

-> Dimension
Window height

-> Dimension
Border width

-> String
Window background color

-> String
Border color

-> String
String color

-> String
String background color

-> [Align]
String Alignments

-> [String]
Strings to be printed

-> [Placement]
Icon Placements

-> [[[Bool]]]
Icons to be printed

-> X ()

Fill a window with a rectangle and a border, and write | a number of strings and a number of icons to given positions

stringToPixel :: (Functor m, MonadIO m) => Display -> String -> m Pixel

Get the Pixel value for a named color: if an invalid name is given the black pixel will be returned.

pixelToString :: MonadIO m => Display -> Pixel -> m String

Convert a Pixel into a String.

fi :: (Integral a, Num b) => a -> b

Short for fromIntegral.